% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOA.DGP1.R
\name{AOA.DGP1}
\alias{AOA.DGP1}
\title{Calculate Area Of Acceptance based on DGP1 simulation which is then used to evaluate the identified IFRs.}
\usage{
AOA.DGP1(Simu, Step, Ci)
}
\arguments{
\item{Simu}{Integer, number of simulations}

\item{Step}{Integer, IRF horizont}

\item{Ci}{Double, between 0 and 1, confidence interval}
}
\value{
A list contains lower and upper bounds arrays of dimension (KxKxh) respectively
}
\description{
Calculate Area Of Acceptance based on DGP1 simulation which is then used to evaluate the identified IFRs.
}
\examples{
NULL
}
