% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{AveragePeriodogram}
\alias{AveragePeriodogram}
\title{Calculate Average periodogram}
\usage{
AveragePeriodogram(
  DataDir,
  Variable,
  Value,
  Species,
  globr,
  reps,
  burnin,
  plotmax = 0,
  Names = 0,
  Title = T,
  yl = T,
  data = F
)
}
\arguments{
\item{DataDir}{Directory containing the data}

\item{Variable}{Name of the Variable in the filename}

\item{Species}{Species to use. Default=SBW}

\item{globr}{Number of Simulations to average}

\item{reps}{Length of simulations}

\item{burnin}{Burnin period}

\item{plotmax}{optional: Ymax height of plot}

\item{Names}{sometimes optional: Vection of names from data file}

\item{Title}{optional: Plot title}

\item{yl}{Remove y label}

\item{data}{optional: Return plot data rather than grob}

\item{Values}{Vector of parameter values for the Moran's I to be calculated for}
}
\value{
Standard Error
}
\description{
Calculate Average periodogram
}
