% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.space.R
\name{B.space}
\alias{B.space}
\title{Calculate a space of potential candidates of decomposition matrix}
\usage{
B.space(B, J = NULL, Rand = NULL)
}
\arguments{
\item{B}{Matrix, (KxK) dimensional original decomposition matrix, usually the cholesky factor}

\item{J}{Integer, number of grid points of rotation angles [0, pi/2]. Function will generates
J + 1 candidates of decomposition matrix}

\item{Rand}{Integer, number of simulated rotation angels from a uniform distribution in the
interval [0, pi].}
}
\value{
An array containing potential candidates of decomposition matrix.
}
\description{
Calculate a space of potential candidates of decomposition matrix
}
