% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.mod.R
\name{cf.mod}
\alias{cf.mod}
\title{Modified Historical Counterfactuals for SVAR Models}
\usage{
cf.mod(x, series = 1, transition = 0, Partial = NULL, Epsname = NULL)
}
\arguments{
\item{x}{my.id}

\item{series}{Integer. indicating the series that should be decomposed.}

\item{transition}{Numeric. Value from [0, 1] indicating how many initial values should be discarded, i.e., 0.1 means that the first 10 per cent observations of the sample are considered as transient.}

\item{Partial}{Integer. indicating which shock is identified}

\item{Epsname}{Characters, Name the structural shocks}
}
\description{
Calculation of historical decomposition (for partial identification) for a proxy identified SVAR object 'my.id'
}
\examples{
NULL
}
