% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covsel.filteralgo.R
\name{covsel.filteralgo}
\alias{covsel.filteralgo}
\title{covsel.filteralgo}
\usage{
covsel.filteralgo(covdata, pa, weights = NULL, force = NULL, corcut = 0.7)
}
\arguments{
\item{covdata}{data.frame containing covariate data extracted at 'pa' locations}

\item{pa}{numeric vector of species presences (1) and absences (0)}

\item{weights}{numeric vector containing the weights for each value in 'pa' (of length 'pa')}

\item{force}{optional character vector indicating the name(s) of the covariate(s) to be forced in the final set}

\item{corcut}{numeric value for the correlation coefficient threshold used for identifying collinearity}
}
\value{
A data.frame of "non-collinear" candidate covariates
}
\description{
Collinearity filtering algorithm
}
\examples{
covdata<-data_covsel$env_vars
dim(covdata)
covdata_filter<-covsel.filteralgo(covdata, pa=data_covsel$pa)
dim(covdata_filter)
}
\author{
Antoine Adde (antoine.adde@unil.ch)
}
