% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{CoxModelSelector}
\alias{CoxModelSelector}
\title{Variable Selection in a Cox Proportional Hazards Model}
\usage{
CoxModelSelector(
  formula,
  data,
  ties = "exact",
  direction = "backwards",
  Alpha = 0.05
)
}
\arguments{
\item{formula}{Formula object with the global model}

\item{data}{Data to be passed to coxph}

\item{ties}{Treatment of tied events. Default="exact'}

\item{direction}{Direction of selection. Default='backwards', options 'backwards'}

\item{Alpha}{P-value threshold. Default = 0.05}
}
\value{
coxph model object of final model selected.
}
\description{
Variable Selection in a Cox Proportional Hazards Model
}
