% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUNctions.R
\name{cut3}
\alias{cut3}
\title{Cut data by interval. Alternative to other options out there,
with slightly nicer looking output (in my humble opinion)}
\usage{
cut3(x, breaks)
}
\arguments{
\item{x}{numeric vector}

\item{breaks}{a numeric vector of break locations}
}
\description{
Cut data by interval. Alternative to other options out there,
with slightly nicer looking output (in my humble opinion)
}

