% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days360.R
\name{days360}
\alias{days360}
\title{days360}
\usage{
days360(start_date, end_date, method = c("US", "EU"))
}
\arguments{
\item{start_date}{Fecha de inicio.}

\item{end_date}{Fecha de fin.}

\item{method}{Método de cálculo de días. Puede ser "US" o "EU". Usa US por default.}
}
\value{
Número de días entre las dos fechas.
}
\description{
Calcula el número de días entre dos fechas utilizando el método 360.
}
\examples{
fecha_inicio <- "2023-01-01"
fecha_fin <- "2024-06-30"
days360(fecha_inicio, fecha_fin)
}
