% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbGetTable}
\alias{dbGetTable}
\title{getTableP}
\usage{
dbGetTable(table, host = "10.192.97.146", port = 5432, dbname = "data")
}
\arguments{
\item{table}{Acá hay que poner alguna. Controla que no esté vacío}

\item{host}{IP del servidor postgresQL}

\item{port}{Puerto del servidor}

\item{dbname}{Base de Datos. Default='data'}
}
\description{
Función para traer una tabla desde el servidor postgresQL
}
