% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbWriteDF}
\alias{dbWriteDF}
\title{dbWriteDF}
\usage{
dbWriteDF(table, df, host = "10.192.97.146", port = 5432, dbname = "data")
}
\description{
Wraper de dbWriteTable del paquete DBI pero que asume las conexiones de mi servidor.
}
\examples{
dbWriteDF(table = 'data', df = dfAGrabar)
}
