% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP1.R
\name{DGP1}
\alias{DGP1}
\title{Baseline DGP from a small dimensional DSGE model calibrated as in Herwartz, Ploedt 2014}
\usage{
DGP1(
  Par,
  Tob,
  Break = NULL,
  Volaregim = NULL,
  VolaRand = NULL,
  Edist = c("tdist", "chidist", "normdist", "unifdist", "gammadist", "betadist",
    "mixed_a", "mixed_b"),
  v = 8,
  Combi = NULL,
  ME = NULL,
  MERand = NULL
)
}
\arguments{
\item{Par}{List, generated by function \code{\link{DSGE2VAR.DGP1}} collects VAR parameters of DSGE}

\item{Tob}{Numeric, length of simulated time series}

\item{Break}{Numeric, time point of structural break (of variances)}

\item{Volaregim}{Vector, a (1x3) or (1x6) vector controlling variance regimes for the structural shocks}

\item{VolaRand}{Vector, a (1x2) or (1x4) vector controlling the min and max of uniform distribution, from which variance regims are drawn}

\item{Edist}{Character, distribution type of simulated structural shocks, Edist = c('tdist', 'chidist', 'normdist', 'unifdist', 'gammadist', 'betadist', 'mixed_a', "mixed_b").}

\item{v}{Numeric, degrees of freedom for distribution of simulated structural shocks}

\item{Combi}{Matrix, a (3x3) matrix, if provided, instruments are linear combination of the true underlying structural shocks}

\item{ME}{Vector, a (1x3) vector containing standard errors of normal white noise as measurement errors in the variables}

\item{MERand}{Vector, a (1x2) vector containing the min and max of uniform distribution, from which the standard errors of normal white noise as measurement errors are drawn}
}
\value{
A list containing simulated macroeconomic variables, structural shocks and their instruments
}
\description{
Baseline DGP from a small dimensional DSGE model calibrated as in Herwartz, Ploedt 2014
}
\examples{
NULL
}
