% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriePBI.R
\name{downloadPBI}
\alias{downloadPBI}
\title{downloadPBI}
\usage{
downloadPBI(
  db = "",
  url = "https://www.indec.gob.ar/ftp/cuadros/economia/sh_oferta_demanda_03_23.xls"
)
}
\arguments{
\item{db}{Utiliza data1 salvo que se le provea otra}

\item{url}{Utiliza https://www.indec.gob.ar/ftp/cuadros/economia/sh_oferta_demanda_03_23.xls por default}
}
\value{
retorna NULL si no hay error. Sino, el error.
}
\description{
Descarga el PBI desde indec. Utiliza la planilla sh_oferta_demanda_03_23.xls
También descarga el EMAE desde "https://www.indec.gob.ar/ftp/cuadros/economia/sh_emae_mensual_base2004.xls"
}
