% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download a Report from REDCap}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable containing the REDCap API token.}

\item{redcapUrl}{The URL of the REDCap API endpoint.}

\item{redcapReportId}{The ID of the report to retrieve from REDCap.}
}
\value{
A tibble containing the data from the specified REDCap report.
}
\description{
This function retrieves a report from a REDCap project using API credentials and returns it as a tibble.
It requires the user to have set an API token in their .Renviron file under the specified name.
}
\examples{
\dontrun{
  result = downloadRedcapReport("MY_REDCAP_TOKEN", "https://redcap.emory.edu/api/", "46524")
  print(result)
}
}
