% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSGE2VAR.DGP1.R
\name{DSGE2VAR.DGP1}
\alias{DSGE2VAR.DGP1}
\title{Solve a 3 dimensional DSGE model and convert it into VAR(2)}
\usage{
DSGE2VAR.DGP1(
  alpha,
  beta,
  k,
  gamma,
  delta_x,
  tau_x,
  tau_pi,
  tau_r,
  rho_x,
  rho_pi,
  rho_r
)
}
\arguments{
\item{rho_r}{}
}
\value{
A list containing VAR and structural parameters.
}
\description{
Solve a 3 dimensional DSGE model and convert it into VAR(2)
}
\examples{
NULL
}
