% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva.best.R
\name{eva.best}
\alias{eva.best}
\title{Test all column ordering and select the best one according to AOA}
\usage{
eva.best(Model, AOA, Step, Bmat, Varname)
}
\arguments{
\item{Model}{A class of 'varest' or 'vec2var'}

\item{AOA}{List generated by function AOA.3d()}

\item{Step}{Integer. Maximum number of iterations for first stage optimization using DEoptim}

\item{Bmat}{A (KxK) impact matrix of covariance matrix of reduced form errors, such that Bmat*Bmat'= Sigma_U / Identification matrix}

\item{Varname}{Character, names of variables in the systems}
}
\value{
A (KxK) impact matrix, which is the best according to AOA
}
\description{
Test all column ordering and select the best one according to AOA
}
\examples{
NULL
}
