% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{eva.RMSE}
\alias{eva.RMSE}
\title{relative mean squared error}
\usage{
eva.RMSE(B, Bhat, xcol = NULL)
}
\arguments{
\item{B}{Matrix of dimension (KxK), the true structural decomposition}

\item{Bhat}{Matrix of dimension (KxK), the estimated decomposition}

\item{xcol}{Numeric, if provided, only the specific column in matrix Bhat is evaluated}
}
\value{
Numeric, relative mean squared error
}
\description{
relative mean squared error
}
\examples{
NULL
}
