% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{eva.sign}
\alias{eva.sign}
\title{check signs / labeling ratio}
\usage{
eva.sign(Bhat, TestSign, Permut = FALSE)
}
\arguments{
\item{Bhat}{Matrix of dimension (KxK), the estimated decomposition}

\item{TestSign}{Matrix of dimension (KxK), containing the true Sign pattern in DGP 1: positive sign; -1: negative sign}

\item{Permut}{Logical, whether apply function to alternative column orderings}
}
\value{
Logical
}
\description{
check signs / labeling ratio
}
\examples{
NULL
}
