% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.AOA.R
\name{Evaluate.AOA}
\alias{Evaluate.AOA}
\title{Evaluate whether the identified IRFs lie in the Area Of Acceptance}
\usage{
Evaluate.AOA(Irf, AOA)
}
\arguments{
\item{Irf}{(KxKxh) array generated by get.my.irf() function. IRF to be evaluated}

\item{AOA}{List generated by AOA.3d or AOA.4d}
}
\value{
a (KxKxh) array containing logical values, TRUE, if the identified IRFs lie in the Area Of Acceptance;
FALSE, otherwise.
}
\description{
Evaluate whether the identified IRFs lie in the Area Of Acceptance
}
\examples{
NULL
}
