% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_legend.R
\name{extract_legend}
\alias{extract_legend}
\title{Extract legends from ggplot}
\usage{
extract_legend(my_ggp)
}
\arguments{
\item{my_ggp}{GGPlot object}
}
\value{
grob containing the legend from my_gpp
}
\description{
Extract legends from ggplot
}
