% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUNctions.R
\name{facFormat}
\alias{facFormat}
\title{Reformat factor levels}
\usage{
facFormat(x, search, replace)
}
\arguments{
\item{x}{factor}

\item{search}{search string of level to rename/combine}

\item{replace}{string to replace level by}
}
\description{
Reformat factor levels
}

