% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_date.R
\name{first_date}
\alias{first_date}
\title{Filtra el primer registro de cada mes}
\usage{
first_date(df, date_col)
}
\arguments{
\item{df}{Dataframe que se desea filtrar.}

\item{date_col}{Columna que contiene las fechas. Debe especificarse sin comillas.}
}
\value{
Un dataframe con únicamente la primera fila de cada mes, basado en la columna de fecha proporcionada.
}
\description{
Esta función recibe un dataframe y una columna de fecha y devuelve un nuevo dataframe que contiene únicamente
la primera fila (según la fecha) de cada mes. Utiliza la sintaxis y funciones del tidyverse junto con evaluación tidy
para especificar la columna de fecha.
}
\examples{
\dontrun{
# Supongamos que 'df' es un dataframe que contiene una columna de fechas llamada 'fecha'.
resultado <- first_date(df, fecha)
}

}
