% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{FirstDefol}
\alias{FirstDefol}
\title{Find the first year with defoliation equal to or greater than some threshold}
\usage{
FirstDefol(x, nyears = 13, Defol = 15, maxyear = 2020)
}
\arguments{
\item{x}{vector including defoliation values.}

\item{nyears}{time span to the data. Default=13}

\item{Defol}{Threshold value. Default=2}

\item{maxyear}{Final year of data collection. Default=2020}
}
\value{
Year of first defoliation
}
\description{
Find the first year with defoliation equal to or greater than some threshold
}
