% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUNctions.R
\name{flexiRound}
\alias{flexiRound}
\title{Rounding a number to a specified base}
\usage{
flexiRound(x, base)
}
\arguments{
\item{x}{a vector of arbitrary length to be rounded}

\item{base}{the base which each element of x will be rounded to}
}
\value{
a numeric vector with elements rounded to the nearest multiple of base
}
\description{
This function rounds a numeric vector input to a user defined base
}
\author{
Philip Goddard
}
\seealso{
\code{round}
}

