% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_1.R
\name{function_1}
\alias{function_1}
\title{Function_1}
\usage{
function_1(x, y)
}
\description{
Your input for this function should be an individual year. Your output should be a data point of interest in the dataset for some player.
}
\examples{
function_1()
}
\keyword{NBA}
