% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_2.R
\name{function_2}
\alias{function_2}
\title{Function_2}
\usage{
function_2(x, y)
}
\description{
Your input for this function should be an individual year. Your output should be the oldest player in the NBA data for that year
}
\examples{
function_2()
}
\keyword{NBA}
