% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_3.R
\name{function_3}
\alias{function_3}
\title{Function_3}
\usage{
function_3(x, y)
}
\description{
Your input for this function should be an individual year. Your function should limit the data for that year to include numeric continuous variables only (not categorical data). Your output should be the correlation matrix for all numeric variables in the data for that year.
}
\examples{
function_3()
}
\keyword{NBA}
