% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-package.R
\docType{package}
\name{functions-package}
\alias{functions}
\alias{functions-package}
\title{functions package}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Ryan Sangston \email{rms2jg@virginia.edu}

}
\keyword{internal}
