% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GearyC.R
\name{GearyC}
\alias{GearyC}
\title{Calculate Geary's C}
\usage{
GearyC(sgrid, data, dclasses)
}
\arguments{
\item{sgrid}{X and Y values of points}

\item{data}{Values of points}

\item{dclasses}{Distance classes}
}
\value{
Weird Data frame with data
}
\description{
Calculate Geary's C
}
