% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.cv.R
\name{get.id.cv}
\alias{get.id.cv}
\title{Cloned from svars: Identification of SVAR models based on Changes in volatility (CV)}
\usage{
get.id.cv(
  Model,
  SB,
  start = NULL,
  end = NULL,
  frequency = NULL,
  format = NULL,
  dateVector = NULL,
  max.iter = 50,
  crit = 0.001,
  restriction_matrix = NULL
)
}
\arguments{
\item{Model}{An object of class 'vars', 'vec2var', 'nlVar'. Estimated VAR object}

\item{SB}{Integer, vector or date character. The structural break is specified either by an integer (number of observations in the pre-break period),
a vector of ts() frequencies if a ts object is used in the VAR or a date character. If a date character is provided, either a date vector containing the whole time line
in the corresponding format (see examples) or common time parameters need to be provided}

\item{start}{Character. Start of the time series (only if dateVector is empty)}

\item{end}{Character. End of the time series (only if dateVector is empty)}

\item{frequency}{Character. Frequency of the time series (only if dateVector is empty)}

\item{format}{Character. Date format (only if dateVector is empty)}

\item{dateVector}{Vector. Vector of time periods containing SB in corresponding format}

\item{restriction_matrix}{Matrix. A matrix containing presupposed entries for matrix B, NA if no restriction is imposed (entries to be estimated). Alternatively, a K^2*K^2 matrix can be passed, where ones on the diagonal designate unrestricted and zeros restricted coefficients. (as suggested in Luetkepohl, 2017, section 5.2.1).}
}
\value{

}
\description{
Cloned from svars: Identification of SVAR models based on Changes in volatility (CV)
}
