% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.cvm.R
\name{get.id.cvm}
\alias{get.id.cvm}
\title{Independence-based identification based on Cramer-von Mises distance}
\usage{
get.id.cvm(Model, dd, Iter, Steptol, Stage2 = NULL, Grid.search = NULL)
}
\arguments{
\item{Model}{A class of 'varest' or 'vec2var'}

\item{dd}{Object of class 'indepTestDist' (generated by 'indepTest' from package 'copula'). A simulated independent sample of the same size as the data.}

\item{Iter}{Integer. Maximum number of iterations for first stage optimization using DEoptim}

\item{Steptol}{Numeric. Tolerance for steps without improvement for DEoptim}

\item{Stage2}{Integer, if provided, second stage optimization will be performed at given iterations}

\item{Grid.search}{}
}
\value{
list object of class "my.id"
}
\description{
Independence-based identification based on Cramer-von Mises distance
}
\examples{
NULL
}
