% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.dcov.R
\name{get.id.dcov}
\alias{get.id.dcov}
\title{Independence-based identification based on distance covariance}
\usage{
get.id.dcov(Model)
}
\arguments{
\item{Model}{A class of 'varest' or 'vec2var'}
}
\value{
list object of class "my.id"
}
\description{
Independence-based identification based on distance covariance
}
\examples{
NULL
}
