% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.iv.R
\name{get.id.iv}
\alias{get.id.iv}
\title{Identification based on Proxy SVAR}
\usage{
get.id.iv(Model, instruments = NULL, Synchro = FALSE)
}
\arguments{
\item{Model}{An object of class 'vars', 'vec2var', 'nlVar'. Estimated VAR object}

\item{instruments}{Vector of dimension (Tx1) containing SINGLE instrument for ONE of structural shocks
with the same length as the time series}

\item{Synchro}{Logical, if Synchro = TRUE, instruments and VAR-residuals are synchronized. This only works if
the given instruments is an object of class "ts".}
}
\value{
list object of class "my.id"
}
\description{
Identification based on Proxy SVAR
}
\examples{
NULL
}
