% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.Sign.R
\name{get.id.Sign}
\alias{get.id.Sign}
\title{Identification based on Sign restrictions}
\usage{
get.id.Sign(
  Model,
  Iter,
  Itermax,
  Signmat,
  RHorizont,
  MT = FALSE,
  Integrate = FALSE,
  PositiveDiag = TRUE,
  Normalize = TRUE,
  Step,
  Ci,
  Plot = FALSE,
  Hist = FALSE,
  Epsname = NULL
)
}
\arguments{
\item{Model}{A class of 'varest' or 'vec2var'}

\item{Iter}{Integer, number of successful draws}

\item{Itermax}{Integer, number of maximum draws}

\item{Signmat}{Matrix, a (KxK) matrix containing Sign restrictions. 1: positive sign; -1: negative sign; NA: unrestricted}

\item{RHorizont}{Integer, RHorizonts of impulse responses, for which the sign restrictions are imposed}

\item{MT}{Logical, if MT = TRUE, Fry and Pagan’s (2011) median target (MT) method will be used}

\item{Integrate}{logical, If Integrate = TRUE, the IRFs will be integrated}

\item{PositiveDiag}{Logical, if PositiveDiag = TRUE, the diagonal of impact matrix will be forced to be positive}

\item{Normalize}{logical, if Normalize = True, size of the shocks will be normalized}

\item{Step}{Integer, IRF horizont}

\item{Ci}{Double, between [0,1] sample quantiles for IRFs}

\item{Plot}{Logical, if Plot = TRUE, IRF plot will be generated}

\item{Hist}{Logical, if Hist = TRUE, histogram of distribution of unrestricted elements in the identified B matrices is provided}

\item{Epsname}{Characters, Name the structural shocks}
}
\value{
A list containing a (KxKxm) array (m-sets) of decomposition matrices which fulfill the sign restrictions, arrays of
impulse responses (median, lower and upper bound)
}
\description{
Identification based on Sign restrictions
}
\examples{
NULL
}
