% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.MBB.fixed.R
\name{get.MBB.fixed}
\alias{get.MBB.fixed}
\title{Moving Block Boostraps with fixed design}
\usage{
get.MBB.fixed(
  My.id,
  design = "fixed",
  b.length = 15,
  Step,
  nboot,
  Ci,
  Integrate = F
)
}
\arguments{
\item{My.id}{An object of of class "my.id", identified SVAR}

\item{design}{character. If design="fixed", a fixed design bootstrap is performed. If design="recursive", a recursive design bootstrap is performed.}

\item{b.length}{Integer, length of each block}

\item{Step}{Integer, Integer, IRF horizons}

\item{nboot}{Integer, number of bootstrap replicates.}

\item{Ci}{Numeric, quantile of boostraped IRFs}

\item{Integrate}{logical, If Integrate = TRUE, the IRFs will be integrated}
}
\value{

}
\description{
Moving Block Boostraps with fixed design
}
\examples{
NULL
}
\references{
Jentsch, Carsten, and Kurt G. Lunsford. 2019. "The Dynamic Effects of Personal and Corporate Income Tax Changes in the United States: Comment." American Economic Review, 109 (7): 2655-78.
}
