% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.my.fevd.R
\name{get.my.fevd}
\alias{get.my.fevd}
\title{Forecast error variance decomposition for SVAR Models}
\usage{
get.my.fevd(My.id = NULL, Step = NULL, Plot = FALSE, Simple = F, IRF = NULL)
}
\arguments{
\item{My.id}{Object of class "my.id", identified SVAR}

\item{Step}{Calculation steps}

\item{Plot}{Logical, plot is generated if Plot = TRUE}

\item{Simple}{Logical, if Simple = TRUE, run only kernel function based on the provided IRF.}

\item{IRF}{Array, of dimension KxKx(Step+1)}
}
\value{
Kx1 list containing FEVD
}
\description{
Forecast error variance decomposition for SVAR Models
}
\examples{
NULL

}
