% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.my.irf.R
\name{get.my.irf}
\alias{get.my.irf}
\title{Return structural impulse response functions (Theta)}
\usage{
get.my.irf(
  Model = NULL,
  Acoef = NULL,
  Bcoef = NULL,
  Simple = FALSE,
  Lag = NULL,
  Step,
  Normalize = FALSE,
  Varname = NULL,
  Bmat = NULL,
  Integrate = FALSE,
  Plot = FALSE
)
}
\arguments{
\item{Model}{A class of 'varest', 'svarest' or 'vec2var'}

\item{Acoef}{A (KxKxp) array. If this is provided, IRF will be calculated based on the given VAR Coefficients.}

\item{Bcoef}{A (Kp + mu) x K dimensional VAR parameter vectors, where mu is the number of deterministic terms}

\item{Simple}{Logical, if Simple = TRUE, run only kernel function. This is particularly useful when function is called within other functions, e.g. bootstraps}

\item{Step}{Integer, IRF horizons}

\item{Normalize}{logical, if Normalize = True, size of the shocks will be normalized}

\item{Varname}{A (1xK) vector of strings, Set variable names manuelly}

\item{Bmat}{A (KxK) decomposition matrix of covariance matrix of reduced form errors, such that Bmat*Bmat'= Sigma_U / Identification matrix}

\item{Integrate}{logical, If Integrate = TRUE, the IRFs will be integrated}

\item{Plot}{logical, If Plot = TRUE, a plot with generic plot function will be generated}
}
\value{
An array collects the structural impulse response functions (Theta)
}
\description{
Return structural impulse response functions (Theta)
}
\examples{
NULL

}
