% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.wild.fixed.Boots.R
\name{get.wild.fixed.Boots}
\alias{get.wild.fixed.Boots}
\title{Wild bootstrap with fixed design}
\usage{
get.wild.fixed.Boots(My.id, Dist = "rademacher", Step, nboot, Ci)
}
\arguments{
\item{My.id}{An object of of class "my.id", identified SVAR}

\item{Dist}{Distribution, from which independent random noises are drawn.}

\item{Step}{Integer, Integer, IRF horizons}

\item{nboot}{Integer, number of bootstrap replicates.}

\item{Ci}{Numeric, quantile of boostraped IRFs}
}
\value{

}
\description{
Wild bootstrap with fixed design
}
