% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDLR.R
\name{getDLR}
\alias{getDLR}
\title{Duplicado con otro nombre para conveniencia
Trae el valor de dolar mep, ccl de BYMA usando
métodos de PPI}
\usage{
getDLR(from = "2014-05-27", to = Sys.Date(), settle = "t+0")
}
\arguments{
\item{from}{Fecha Inicio. Mínima 2015-05-27}

\item{to}{hasta donde}
}
\value{
Una tibble con la serie. Anula la devolución del gráfico como
segundo componente del valor retornado según viene de getPPIDLR
}
\description{
Duplicado con otro nombre para conveniencia
Trae el valor de dolar mep, ccl de BYMA usando
métodos de PPI
}
\examples{
getDLR(from = "2020-09-14", to = Suys.Date(), settle = "t+2") -> Devuelve la tibble

}
