% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriePBI.R
\name{getEMAE}
\alias{getEMAE}
\title{getEMAE}
\usage{
getEMAE(db = "", download = FALSE)
}
\arguments{
\item{db}{Indicar base. Por default data1.}

\item{download}{FALSE. Si TRUE llama functions::downloadPBI}
}
\value{
df con el EMAE.
}
\description{
Devuelve el EMAE almacenado en la base.
}
