% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriePBI.R
\name{getPBI}
\alias{getPBI}
\title{getPBI}
\usage{
getPBI(table = "pbiCorriente", format = "T", db = "", download = FALSE)
}
\arguments{
\item{table}{Que tabla retornar. pbiCorriente, pbiConstante.}

\item{format}{T para Trimestral. A para Anual}

\item{db}{Por default busca en data1. Sino indicar.}

\item{download}{Por default es FALSE. Si es TRUE llama a functions::downloadPBI para actualizar.}
}
\value{
df con el PBI
}
\description{
Retorna el PBI que trae desde la base de datos db.
}
