% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggMultiplot}
\alias{ggMultiplot}
\title{Make a multiplot from a list containing ggplots}
\usage{
ggMultiplot(plots = NULL, cols = 1)
}
\arguments{
\item{plots}{a list containing a plot in each element}

\item{cols}{number of columns to be produced in the multiplot}
}
\description{
Make a multiplot from a list containing ggplots
}
\author{
Philip Goddard, modified from an R blog I once found
}

