% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUNctions.R
\name{ggplotListBar}
\alias{ggplotListBar}
\title{Make a list of stacked bar charts plots using ggplot}
\usage{
ggplotListBar(df, outcome)
}
\arguments{
\item{df}{a data frame containing categorical data}

\item{outcome}{the groups of the data}
}
\description{
Make a list of stacked bar charts plots using ggplot
}
\author{
Philip Goddard
}

