% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplotListDens}
\alias{ggplotListDens}
\title{Make a list of density plots using ggplot}
\usage{
ggplotListDens(df, outcome)
}
\arguments{
\item{df}{a data frame containing numerical data}

\item{outcome}{the groups of the data}
}
\description{
Make a list of density plots using ggplot
}
\author{
Philip Goddard
}

