% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplotListHist}
\alias{ggplotListHist}
\title{Make a list of stacked histograms plots using ggplot}
\usage{
ggplotListHist(df, outcome)
}
\arguments{
\item{df}{a data frame containing count data}

\item{outcome}{the groups of the data}
}
\description{
Make a list of stacked histograms plots using ggplot
}
\author{
Philip Goddard
}

