% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givens.Q.R
\name{givens.Q}
\alias{givens.Q}
\title{Calculate (a sequence of) Givens Rotation matrices}
\usage{
givens.Q(theta, K)
}
\arguments{
\item{theta}{Vector, containing \code{choose(K,2)} rotation angles}

\item{K}{Integer, dimension of decomposition matrix}
}
\value{
a (KxK) rotation matrix
}
\description{
Calculate (a sequence of) Givens Rotation matrices
}
\examples{
NULL
}
