% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.sum.R
\name{glm.sum}
\alias{glm.sum}
\title{glm summary function}
\usage{
glm.sum(m, dispersion = T)
}
\arguments{
\item{m}{glm model}

\item{dispersion}{numeric value for dispersion (not used)}
}
\description{
Produces Genstat like summary information for a glm
}
\examples{
DT <- data.table(dv=c(rnorm(100,10,2),rnorm(100,20,2)),iv=as.factor(c(rep("A",100),rep("B",100))))
varcheck(DT,"dv","iv")
}
