% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabaTabla.R
\name{grabaTabla}
\alias{grabaTabla}
\title{grabaTabla}
\usage{
grabaTabla(
  fecha = Sys.Date(),
  variable,
  path = "~/OneDrive/outlier/docs/tablasGraficos",
  vwidth,
  vheight
)
}
\arguments{
\item{fecha}{Fecha de la tabla}

\item{variable}{Tabla a grabar}

\item{path}{Ruta donde se grabará la tabla}

\item{vwidth}{Ancho de la tabla}

\item{vheight}{Alto de la tabla}
}
\description{
Graba una tabla en formato png
}
\examples{
grabaTabla(variable = tablaMLC)
grabaTabla(variable = tablaMLC, vwidth = 990, vheight = 409)
}
