% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_icon.R
\name{gradient_icon}
\alias{gradient_icon}
\title{gradient_icon}
\usage{
gradient_icon(col = colorRampPalette(c("grey90", "grey20"))(100),
  labels = c(0, 1), tdist = 0.01, bars = 10, title = NULL,
  dist = c(0.15, 0.15), tcol = "white")
}
\arguments{
\item{col}{spatialpointsdataframe with the parameters}

\item{labels}{minimum and maximum value}

\item{tdist}{distance of the labels from the gradient scale}

\item{bars}{number of lines in the gradient scale}

\item{title}{title of the gradient scale}

\item{dist}{distance from x and y top corner}

\item{tcol}{color of labels and title}
}
\description{
Generates a gradient scale
}
\examples{
export(gradient_icon(col=colorRampPalette(c('grey90','grey20'))(100), labels = c(500,1000), tdist=0.05,bars=10,title=NULL),res2=30,ratio=0.8,'gradient_icon',type='png') 
}
\keyword{gradient,}
\keyword{plot}
