% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.mod.R
\name{hd.mod}
\alias{hd.mod}
\title{modifited historical decomposition}
\usage{
hd.mod(x, series = 1, Partial = NULL, Epsname = NULL)
}
\arguments{
\item{x}{svar or my.id}

\item{series}{which series to be decomposed}

\item{Partial}{partial identification, by default Partial = NULL}

\item{Epsname}{Vector of shock names (string)}
}
\value{
data.frame
}
\description{
modifited historical decomposition
}
\examples{
NULL
}
