% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoshenkopelman.R
\name{hosen}
\alias{hosen}
\title{Identifies contiguous patches in the landscape}
\usage{
hosen(squared, mincd = 0)
}
\arguments{
\item{squared}{Snapshot of landscape output}
}
\value{
Matrix of labelled patches
}
\description{
Identifies contiguous patches in the landscape
}
