% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon}
\alias{icon}
\title{icon}
\usage{
icon(pch = 24, col = "black", bg = "white", filename = NULL)
}
\arguments{
\item{pch}{symbol to use}

\item{col}{color border}

\item{bg}{color background}

\item{filename}{the name of the icon when saving as a png, if NULL the png is not exported}
}
\description{
generate an icon
}
\examples{
# generate and export an icon
icon(pch=21, col='black',bg='white',filename='icon')
}
\keyword{export,}
\keyword{icon,}
\keyword{png}
