% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{indextowavelength}
\alias{indextowavelength}
\title{\code{indextowavelength} converts an index into a wavelength for the terraref dataset.}
\usage{
indextowavelength(index)
}
\arguments{
\item{index}{A number}
}
\value{
wavelenght (nm)
}
\description{
\code{indextowavelength} converts an index into a wavelength for the terraref dataset.
}
\examples{
indextowavelength(450)
}
