% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUNctions.R
\name{insert_col}
\alias{insert_col}
\title{Insert a column into a data frame in  a specified position}
\usage{
insert_col(df, new_col, col_name = "new_col", numeric_index = TRUE,
  insert_index, insert_name, insert_left = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{new_col}{a vector of same length of df to be added to df}

\item{col_name}{name of the new column}

\item{numeric_index}{should the new column be inserted by index (TRUE) or name (FALSE)}

\item{insert_index}{position for new column to be inserted}

\item{insert_name}{if numeric_index is FALSE, the column name the new column will be inserted adjacent to}

\item{insert_left}{if numeric_index is FALSE, will
the column be inserted to left of insert_name (TRUE), or to the right (FALSE)}
}
\description{
Insert a column into a data frame in  a specified position
}
\author{
Philip Goddard
}

