% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUNctions.R
\name{is.binary}
\alias{is.binary}
\title{Determine if vector is binary}
\usage{
is.binary(x)
}
\arguments{
\item{x}{input vector}
}
\description{
Predicate function for testing only 0, 1 in vector
}
\author{
Philip Goddard
}

