% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml2kmz.R
\name{kml2kmz}
\alias{kml2kmz}
\title{kml2kmz}
\usage{
kml2kmz(files_to_zip = c("icon.jpg", "track_kml.kml"),
  filename = "track_kmz")
}
\arguments{
\item{files_to_zip}{a vector with the files that need to be zipped into the kmz file}

\item{filename}{output file name of the kmz}
}
\description{
Generates a kmz from a klm
}
\examples{
## Example using png a locally stored png
# Prepare a dataframe 
obj <- data.frame(altitude=1:100,colour=heat.colors(100),heading=1:100,variable=runif(100),longitude=seq(15.202020, 16.202020,length.out=100), latitude=seq(45.459069,46.543523,length.out=100))
coordinates(obj) <- c('longitude','latitude')
proj4string(obj)<- CRS("+init=epsg:4326")

# First generate and save icon
icon(pch=24, col='black',bg='white',filename='icon')

# Use function with the generated icon
kml2(obj,altitude='altitude',colour='colour',heading='heading',size='variable',icon="icon.png",filename='track_kml.kml')
kml2kmz(files_to_zip=c('icon.png','track_kml.kml'),filename='track_kmz')
}
\keyword{kml,kmz}
