% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthLabeller.r
\name{labelNLMeses}
\alias{labelNLMeses}
\title{labelNLMeses
Es un una función que se usa dentro de scale_x_date y similares
y sirve para convertir los meses del año en español,
y además les pone un separador. En el caso de la variante
slash, el separador es "-" y en el caso de la variante
NL el separador es un salto de línea.}
\usage{
labelNLMeses(date)
}
\description{
#' @param date Un objeto de clase Date que es pasado automáticamente
por ggplot2.
#' @return el vector con las etiquetas de los meses en español y el separador
}
