% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthLabeller.r
\name{labelSlashMeses}
\alias{labelSlashMeses}
\title{labelSlashMeses}
\usage{
labelSlashMeses(date)
}
\arguments{
\item{date}{Un objeto de clase Date que es pasado automáticamente
por ggplot2.}
}
\value{
el vector con las etiquetas de los meses en español y el separador
}
\description{
Es un una función que se usa dentro de scale_x_date y similares
y sirve para convertir los meses del año en español,
y además les pone un separador. En el caso de la variante
slash, el separador es "-" y en el caso de la variante
NL el separador es un salto de línea.
}
