% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayFunctions.R
\name{LandscapeDataToArray}
\alias{LandscapeDataToArray}
\title{Convert Cellular model output to 3 dimensional array}
\usage{
LandscapeDataToArray(d, verbose = F)
}
\arguments{
\item{data}{Output file from Cellular Automaton (OutbreakLandscape)}
}
\value{
A 3-D Array
}
\description{
Convert Cellular model output to 3 dimensional array
}
