% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_date.R
\name{las_date}
\alias{las_date}
\title{Filtra el último registro de cada mes.}
\usage{
las_date(df, date_col)
}
\arguments{
\item{df}{Dataframe que se desea filtrar.}

\item{date_col}{Columna que contiene las fechas. Debe especificarse sin comillas.}
}
\value{
Un dataframe con únicamente la última fila de cada mes, basado en la columna de fecha proporcionada.
}
\description{
Esta función recibe un dataframe y una columna de fecha y devuelve un nuevo dataframe que contiene únicamente
la última fila (según la fecha) de cada mes. Se utiliza el enfoque de tidyverse y evaluación tidy para
especificar la columna de fecha.
Es ideal para filtrar datos de series temporales, donde se desea conservar únicamente el último registro
de la serie por tratarse de stocks.
}
\examples{
\dontrun{
# Supongamos que 'df' es un dataframe que contiene una columna de fechas llamada 'fecha'.
resultado <- filter_last_date(df, fecha)
}

}
