% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_msg.R
\name{log_msg}
\alias{log_msg}
\title{Loguea un mensaje en un archivo de log con un timestamp y nivel de severidad.}
\usage{
log_msg(msg, level = "INFO", log_file = "app.log")
}
\arguments{
\item{msg}{El mensaje a loguear.}

\item{level}{El nivel de severidad del mensaje (por defecto es "INFO").}

\item{log_file}{El archivo donde se guardará el log (por defecto es "app.log").}
}
\value{
No retorna ningún valor.
}
\description{
Loguea un mensaje en un archivo de log con un timestamp y nivel de severidad.
}
\details{
Esta función escribe un mensaje en un archivo de log especificado,
incluyendo un timestamp y un nivel de severidad (INFO, WARNING, ERROR).
}
\examples{
log_msg("Script finalizado correctamente.", "INFO")

}
