% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{ma}
\alias{ma}
\title{Calculate Moving Average}
\usage{
ma(x, n = 5)
}
\arguments{
\item{x}{Vector to calculate ma from}

\item{n}{window for moving average}
}
\value{
Moving average
}
\description{
Calculate Moving Average
}
