% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalMeans.R
\name{Means}
\alias{Means}
\title{Summary statistics for global temporal data (Output...)}
\usage{
Means(workingdir, vari, i, burnin, reps = 0, r = 32, fun = "mean")
}
\arguments{
\item{workingdir}{Directory the data is in}

\item{vari}{Variable}

\item{i}{Value}

\item{burnin}{Length of burn in period}

\item{reps}{Total length of data. Default=nrow(data).}

\item{r}{Total number of simulations. Default=32.}

\item{fun}{Function. Default="mean".}
}
\value{
Summary statistics
}
\description{
Returns summary statistics for outputs of the model from Output... files
}
