% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanSize.R
\name{MeanSize}
\alias{MeanSize}
\title{Generate summary statistics for OutbreakLandscape file}
\usage{
MeanSize(spacedata, s = 0)
}
\arguments{
\item{spacedata}{OutbreakLandscape file}
}
\value{
Summary Statistics table
}
\description{
Generate summary statistics for OutbreakLandscape file
}
