% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoranI.R
\name{MoranI}
\alias{MoranI}
\title{Calculate Moran's I}
\usage{
MoranI(sgrid, data, dclasses, weight = F)
}
\arguments{
\item{sgrid}{X and Y values of points}

\item{data}{Values of points}

\item{dclasses}{Distance classes}
}
\value{
Weird Data frame with data
}
\description{
Calculate Moran's I
}
