% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{Mortality}
\alias{Mortality}
\title{Find the first year with cumulative defoliation equal to or greater than some threshold}
\usage{
Mortality(x, nyears = 13, maxyear = 2020, MortDefol = 15)
}
\arguments{
\item{x}{vector including defoliation values.}

\item{nyears}{time span to the data. Default=13}

\item{maxyear}{Final year of data collection. Default=2020}

\item{MortDefol}{Threshold value. Default=15}
}
\value{
Year of mortality
}
\description{
Find the first year with cumulative defoliation equal to or greater than some threshold
}
