% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{mround}
\alias{mround}
\title{Round}
\usage{
mround(x, base)

mround(x, base)
}
\arguments{
\item{x}{Ko te tau mo whakaawhiwhi.}

\item{base}{Ko te putake}
}
\value{
Rounded value(s) of x

Rounded value(s) of x
}
\description{
Round

Whakaawhiwhi
}
