% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.bigextract.R
\name{nsdm.bigextract}
\alias{nsdm.bigextract}
\title{nsdm.bigextract}
\usage{
nsdm.bigextract(
  cov,
  data,
  rst_ref,
  cov_info,
  t_match = FALSE,
  tmatch_scheme = "npts",
  p_int = NULL,
  ex_pint = TRUE,
  nzvt = 10,
  nsplits = ncores
)
}
\arguments{
\item{cov}{A character vector indicating paths for each individual raster layer}

\item{data}{The nsdm.pseudoabsences object to be used for covariate extraction}

\item{rst_ref}{Raster used as reference (cell identifiers, crs, extent etc.)}

\item{cov_info}{Data.frame containing key information on covariate, as generated by nsdm.covinfo}

\item{t_match}{Logical for doing temporal matching (TRUE) or not (FALSE)}

\item{tmatch_scheme}{A character string for the name of temporal matching scheme considered ("npts"=nearest previous time slice or "nts"=nearest time slice)}

\item{p_int}{Numeric interval for identifying period interval (e.g. 1981:2010 for historical climate normals)}

\item{ex_pint}{Logical for considering period interval covariates as static (TRUE) or not (FALSE)}

\item{nzvt}{Numeric threshold for the minimal number of unique values for a covariate to bo kept in the candidate set}

\item{nsplits}{Number of cores to be used for parallel extraction}
}
\value{
An updated nsdm.pseudoabsences object
}
\description{
Parallel loading and spatiotemporal extraction of large raster sets
}
\author{
Antoine Adde (aadde@unil.ch)
}
