% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.ceval.R
\name{nsdm.ceval}
\alias{nsdm.ceval}
\title{Model evaluation core function}
\usage{
nsdm.ceval(f, pa, tesdat, crit, tre = numeric())
}
\description{
Model evaluation core function
}
\author{
Philipp Brun (philipp.brun@wsl.ch) and Antoine Adde (aadde@unil.ch)
}
