% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.covfilter.R
\name{nsdm.covfilter}
\alias{nsdm.covfilter}
\title{nsdm.covfilter}
\usage{
nsdm.covfilter(covdata, pa, method, weights, force = NULL, corcut = 0)
}
\arguments{
\item{covdata}{A data.frame containing covariate data}

\item{pa}{A vector of presences (1) and absences (0)}

\item{method}{A character string indicating the target method for covariate ranking (default=quadratic glm p-values)}

\item{weights}{A vector of weights (of length pa)}

\item{force}{A character vector indicating the name(s) of covariates to be forced in final set (e.g.: mainEU)}

\item{corcut}{The value (numeric) of the correlation coefficient threshold used for identifying colinearity}
}
\value{
A thinned set of "non-colinear" candidate covariates
}
\description{
Colinearity filtering algorithm function
}
\author{
Antoine Adde (aadde@unil.ch)
}
