% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.covinfo.R
\name{nsdm.covinfo}
\alias{nsdm.covinfo}
\title{nsdm.covinfo}
\usage{
nsdm.covinfo(cov_path, save_path, time_cov, focal_cov)
}
\arguments{
\item{cov_path}{A character string for the root path where covariates are stored}

\item{save_path}{A character string for the path to which write the .xlsx covariate information table}

\item{dyna_cov}{A character vector containing the names of datasets equipped with temporally dynamic covariates}

\item{focal_cat}{A character vector containing the names of categories equipped with temporally dynamic covariates}
}
\value{
An xlsx covariate information table
}
\description{
Generate covariate information table based on covariate paths/names
}
\author{
Antoine Adde (aadde@unil.ch)
}
