% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.covselrk.R
\name{nsdm.covselrk}
\alias{nsdm.covselrk}
\title{nsdm.covselrk}
\usage{
nsdm.covselrk(embed, species_name)
}
\arguments{
\item{embed}{An nsdm.embedsel output}

\item{species_name}{A character string indicating the name of the target taxon}
}
\value{
A ranked list of candidate covariates after selection procedure
}
\description{
Ranking of covariates after selection procedure
}
\author{
Antoine Adde (aadde@unil.ch)
}
