% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.covsub.R
\name{nsdm.covsub}
\alias{nsdm.covsub}
\title{nsdm.covsub}
\usage{
nsdm.covsub(
  covdata,
  rasterdata,
  ranks,
  thre,
  max.thre,
  glo.out = NULL,
  glo.xy = NULL,
  glo.buff = NULL
)
}
\arguments{
\item{covdata}{A data.frame containing covariate data}

\item{rasterdata}{A raster stack containing corresponding original covdata}

\item{ranks}{A nsdm.covselrk output}

\item{thre}{The target number (numeric) of covariates in a model}

\item{max.thre}{The maximum possible number (numeric) of covariates in a model}

\item{glo.out}{If mainGLO included as covariate, GLO model output}

\item{glo.xy}{If mainGLO included as covariate, two column matrix of for XY GLO occurences coordinates}

\item{glo.buff}{If mainGLO included as covariate, buffer radius used for mainGLO extraction}
}
\description{
Final subsetting of covariate set after selection and ranking procedures
}
\author{
Antoine Adde
}
