% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.df_or_rast.R
\name{nsdm.df_or_rast}
\alias{nsdm.df_or_rast}
\title{Predict to data.frame or raster}
\usage{
nsdm.df_or_rast(mod, nwdat, ...)
}
\description{
Predict to data.frame or raster
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
