% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.disaggregate.R
\name{nsdm.disaggregate}
\alias{nsdm.disaggregate}
\title{nsdm.disaggregate}
\usage{
nsdm.disaggregate(
  pres = numeric(),
  rst,
  thindist = 0,
  thinyear = 0,
  max_uncertain = 9999,
  min_occ = 0,
  ncores = 1
)
}
\arguments{
\item{pres}{Data.frame containing species occurence data with key columns "species" (taxon name), "year" (observation year), "uncertainty" (coordinate uncertainty)}

\item{rst}{Reference raster used in subsequent analyses}

\item{thindist}{Distance (numeric; in rst units) used for spatial disaggregation (minimal distance between two occurences)}

\item{thinyear}{Time period (numeric; in years) used for temporal disaggregation (minimal number of years between two occurences at same pixel)}

\item{max_uncertain}{Maximum coordinate uncertainty (numeric; in "uncertainty unit) for an occurence to be retained in the modelling set}

\item{min_occ}{Minimal number of occurences (numeric) required for a species to be modelled}

\item{ncores}{Number of cores (numeric) used during parallel operations}
}
\value{
An updated Spatial point dataframe object with disaggregated species occurence data
}
\description{
Spatiotemporal disaggregation of species data
}
\author{
Antoine Adde (aadde@unil.ch)
}
