% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.embedsel.R
\name{nsdm.embedsel}
\alias{nsdm.embedsel}
\title{nsdm.embedsel}
\usage{
nsdm.embedsel(covdata, pa, weights, force = NULL, nthreads = detectCores()/2)
}
\arguments{
\item{covdata}{A data.frame containing covariate data}

\item{pa}{A vector of presences (1) and absences (0)}

\item{weights}{Vector of weights of length pa}

\item{force}{A character vector indicating the name(s) of covariates to be forced in final set (e.g.: mainEU)}

\item{nthreads}{Number (numeric) of cores to be used during parallel operations}
}
\value{
A ranked list of covariates selected after the embedded procedure
}
\description{
Embedded covariate selection
}
\author{
Antoine Adde (aadde@unil.ch)
}
