% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.fastraster.R
\name{nsdm.fastraster}
\alias{nsdm.fastraster}
\title{nsdm.fastraster}
\usage{
nsdm.fastraster(files, nsplits = parallel::detectCores() - 1)
}
\arguments{
\item{files}{A character vector containing paths to individual raster layers}

\item{nsplits}{Number of cores (numeric) to be used during parallel loading}
}
\value{
A raster stack
}
\description{
Parallel loading of raster data
}
\author{
Antoine Adde (aadde@unil.ch)
}
