% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.filtersel.R
\name{nsdm.filtersel}
\alias{nsdm.filtersel}
\title{nsdm.filtersel}
\usage{
nsdm.filtersel(
  covdata,
  pa,
  datasets,
  varnames = NULL,
  weights = NULL,
  focals = NULL,
  method,
  force = NULL,
  corcut = 0
)
}
\arguments{
\item{covdata}{A data.frame containing covariate data}

\item{pa}{A vector of presences (1) and absences (0)}

\item{datasets}{A character vector indicating the names of the datasets (for each candidate covariate)}

\item{varnames}{A character vector indicating the names of the variables. Only useful if one variable is supplied with multiple attributes (e.g. min, max, mean) and focal sizes}

\item{weights}{A numeric vector of weights of length pa}

\item{focals}{A character vector indicating which datasets include focals}

\item{method}{A character string indicating the target method for covariate ranking (default=quadratic glm p-values)}

\item{force}{A character vector indicating the name of the covariates to be forced in the final set (e.g.: mainEU)}

\item{corcut}{The value (numeric) of the correlation coefficient threshold for identifying colinearity}
}
\value{
A thinned set of "non-colinear" candidate covariates
}
\description{
Apply the colinearity filtering algorithm at each target level (i=covariates with focals; ii=within each category; iii= all remainders)
}
\author{
Antoine Adde (aadde@unil.ch)
}
