% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.fitting2.R
\name{nsdm.flex2}
\alias{nsdm.flex2}
\title{nsdm.flex2}
\usage{
nsdm.flex2(
  x = numeric(),
  pa = numeric(),
  env_vars = data.frame(),
  taxon = character(),
  replicatetype = character(),
  reps,
  strata = NA,
  mod_args = list(),
  level,
  ncores,
  save = FALSE,
  project = NA,
  path = NA,
  tmp_path
)
}
\arguments{
\item{x}{A nsdm.pseudoabsences object}

\item{pa}{Vector with presence/absence values}

\item{env_vars}{Data.frame with environmental predictors}

\item{taxon}{A character sting of the taxon for which models are fitted}

\item{replicatetype}{(how) should replicates be generated? may be 'none', 'splitsample',
'cv' 'block-cv'}

\item{reps}{Number of replicates}

\item{strata}{a numeric vector of the same length as observations with integers separating
cross validation replicates (used when replicatetype='block-cv')}

\item{mod_args}{List with elements of class 'multi.input' which specify models to be fitted}

\item{level}{Character indicating the level evaluated (e.g. glo or loc)}

\item{ncores}{Number of cores to be used during parallel operations}

\item{save}{should the model be saved in a structured way? (not implemented yet)}

\item{project}{character indicating the name of the project within which the models are run
(later used to define saving directories)}

\item{path}{where to save? (not implemented yet)}

\item{tmp_path}{Character indicating the path where to store temporary outputs}
}
\value{
A nsdm.fit object including slots for meta info, testing data for evaluation, and model objects
}
\description{
Model fitting (parallel version)
}
\author{
Philipp Brun (philipp.brun@wsl.ch) and Antoine Adde (aadde@unil.ch)
}
