% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.map.R
\name{nsdm.map}
\alias{nsdm.map}
\title{nsdm.map}
\usage{
nsdm.map(
  template,
  nona_ix,
  pred,
  species_name,
  model_name,
  level = NA,
  nesting_name = NA,
  rcp_name = NA,
  period_name = NA
)
}
\arguments{
\item{template}{An empty raster layer to fill with predicted values}

\item{nona_ix}{Numeric indices of cells that should be filled with predicted values}

\item{pred}{List containing vectors of fitted values (one list by modelling algorithm)}

\item{species_name}{A character string indicating the name of the taxon for which models are fitted}

\item{model_name}{A character vector indicating the name(s) of the modelling algorithm(s) for which predicted values are provided}
}
\value{
A list of rasters filled with predicted values
}
\description{
Fill empty template rasters with predicted values
}
\author{
Antoine Adde (aadde@unil.ch)
}
