% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.prd.R
\name{nsdm.prd}
\alias{nsdm.prd}
\title{Feed the predict functions depending on model type}
\usage{
nsdm.prd(mod, tst)
}
\description{
Feed the predict functions depending on model type
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
