% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.predict.R
\name{nsdm.predict}
\alias{nsdm.predict}
\title{nsdm.predict}
\usage{
nsdm.predict(models, nwdata, nsplits = NULL)
}
\arguments{
\item{models}{A nsdm.flex object containing fitted models}

\item{nwdata}{A data.frame object containing covariate data for predictions}

\item{nsplits}{Number of splits for parallel operations}
}
\value{
A vector of predicted values
}
\description{
Compute predictions (parallel) for each target modelling algorithm
}
\author{
Antoine Adde (aadde@unil.ch)
}
