% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.respcurve.R
\name{nsdm.respcurve}
\alias{nsdm.respcurve}
\title{nsdm.respcurve}
\usage{
nsdm.respcurve(
  models,
  Data,
  scaleparam = NULL,
  factor = 1,
  model_name,
  species_name,
  plotting = TRUE,
  save_path = NULL,
  ncores = 1
)
}
\arguments{
\item{models}{A nsdm.fit object containing fitted model(s)}

\item{Data}{A data.frame object containing covariate values used for model fitting}

\item{scaleparam}{Attribute part of a scaled matrix containing the scaling parameters to be reapplied}

\item{factor}{When applicable, the value by which original covariates have been multiplied (e.g. by 10, 100, etc.) for storage purposes}

\item{model_name}{A character string indicating the name of the target modelling algorithm}

\item{species_name}{A character string indicating the name of the target taxon}

\item{plotting}{Logical. Plot response curves?}

\item{save_path}{A character string indicating the path where the plots will be saved}

\item{ncores}{Number of cores to be used during parallel operations}
}
\value{
A list containing the values of the response curves for each individual covariate and model
}
\description{
Covariate response curves computation
}
\author{
Antoine Adde (aadde@unil.ch)
}
