% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.retrieve4pred.R
\name{nsdm.retrieve4pred}
\alias{nsdm.retrieve4pred}
\title{nsdm.retrieve4pred}
\usage{
nsdm.retrieve4pred(
  covstk,
  scaleparam,
  mask = NULL,
  observational = NULL,
  obsval = NULL
)
}
\arguments{
\item{covstk}{A raster stack containing the covariate layers used for prediction}

\item{scaleparam}{Attribute part of a scaled matrix containing the scaling parameters to be reapplied}

\item{mask}{A character string for the path to .rds file containing the raster with cells to be masked (value of 0) when predicting}

\item{observational}{A character vector containing the names of "observational" covariates for which the same value (obsval) will be used for prediction}

\item{obsval}{A character string indicating the strategy used to set the common value for observational covariates ("zero", "median", or "max"imum)}
}
\value{
A data.frame object on which predictions will be applied and a companion vector containing the indices for non-na cells on the original raster stack
}
\description{
Prepare covariate data for prediction
}
\author{
Antoine Adde (aadde@unil.ch)
}
