% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.setparam.R
\name{nsdm.setparam}
\alias{nsdm.setparam}
\title{nsdm.setparam}
\usage{
nsdm.setparam(
  model_name,
  param_grid,
  covariate_names,
  tmp_path,
  ncov.esm = NULL,
  comb.esm = NULL,
  force.esm = NULL,
  weights = 1,
  nthreads
)
}
\arguments{
\item{model_name}{A character string indicating the name of the target modelling algorithm}

\item{param_grid}{A character string indicating the complete path where the .xlsx document containing the parameter settings is stored}

\item{covariate_names}{A character vector indicating covariate names}

\item{tmp_path}{A character string indicating the path where to store temporary outputs}

\item{ncov.esm}{If the ESM framework is used, a numeric value indicating the total number of covariates evaluated}

\item{comb.esm}{If the ESM framework is used, a numeric value indicating the number of covariates combined in each individual small model}

\item{force.esm}{If the ESM framework is used, a character string indicating the name of covariate(s) to be forced (typically mainGLO)}

\item{weights}{A numeric vector of weights to be applied}

\item{nthreads}{Number (numeric) of cores to be used during parallel operations}
}
\value{
List with elements of class 'multi.input' which specify models to be fitted
}
\description{
Define model parameter settings
}
\author{
Antoine Adde (aadde@unil.ch)
}
