% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.stratified.R
\name{nsdm.stratified}
\alias{nsdm.stratified}
\title{nsdm.stratified}
\usage{
nsdm.stratified(
  df,
  group,
  size,
  select = NULL,
  replace = FALSE,
  bothSets = FALSE
)
}
\arguments{
\item{df}{The input data.frame}

\item{group}{A character vector of the column or columns that make up the "strata".}

\item{size}{The desired sample size}

\item{select}{For sampling with replacement}

\item{replace}{This allows you to subset the groups in the sampling process. This is a list. For instance, if your group variable was "Group", and it contained three strata, "A", "B", and "C", but you only wanted to sample from "A" and "C", you can use select = list(Group = c("A", "C"))}

\item{bothSets}{}
}
\description{
Stratified sampling
}
\author{
Ananda Mahto (mrdwab@gmail.com) and Antoine Adde (aadde@unil.ch)
}
