% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayFunctions.R
\name{nslice}
\alias{nslice}
\title{Return number of slices in a 3D array}
\usage{
nslice(x)
}
\arguments{
\item{x}{Three-Dimensional Array}
}
\value{
The number of slices in the array
}
\description{
Return number of slices in a 3D array
}
