% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset.R
\name{offset}
\alias{offset}
\title{Offset}
\usage{
offset(df, direction = 1, names = 2, pivot = 100)
}
\arguments{
\item{df}{Dataframe}

\item{direction}{1 hacia adelante. -1 hacia atras}

\item{names}{Si es long es el numero de columna donde están los nombres}

\item{pivot}{valor base para el cálculo del offset. Por defecto 100.}
}
\value{
Dataframe con las columnas adicionales
}
\description{
Calcula la variación a partir del primer elemento del df para cada uno de los datos.
Trabaja con formato wider o longer.
Si es longer, default, espera en names el número de columan donde esta los nombres
direction es para ver si lo hace para adelante o para atras (1 adelante, -1 hacia atrás)
ideal para pasarle un df con lo que queremos comparar y ver como fue la evolución de cada una respecto al ppio.
}
\examples{
\dontrun{
offset(df, direction = 1)
offset(df, direction = -1, names = 2, pivot = 45)
offset(df, direction = 1, names = 0)
}
}
