% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_plot_resid_dense.R
\name{ols_plot_resid_dense}
\alias{ols_plot_resid_dense}
\title{Residual density plot}
\usage{
ols_plot_resid_dense(model)
}
\arguments{
\item{model}{an lm model 
scores.}
}
\value{
A ggplot density plot.
}
\description{
ols_plot_resid_dense plots a residual density plot.
}
\details{
This function is an adjunct to olsrr package which adds density plot functionality.
}
\examples{
DF <- data.frame(x=rnorm(30,10,2),
				Treat=as.factor(c(rep("A",10),rep("B","10"),rep("C","10"))),
				Block=as.factor(rep(c(1,2),15)))
model=lm(x~Block+Treat,DF)
ols_plot_resid_dense(model)
}
