% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Calculate the approximation to a data x based on PCs}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix of data to be approximated.}

\item{npc}{The number of principal components to use in the approximation.}
}
\value{
A numeric matrix approximated using the specified number of principal components.
}
\description{
This function returns an approximation to the data x based on
npc PCs. The function also reconstructs the data, ensuring that it is
rescaled and centered to match the original dataset.
}
\examples{
set.seed(123)
data_matrix = matrix(rnorm(100), nrow=10)
approx_data = pcApprox(data_matrix, npc=2)
print(approx_data)
}
