% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Calculate periodogram}
\usage{
periodogram(sbw, smoothing = 1, spli = 2, val = "", averageline = 0, data = F)
}
\arguments{
\item{sbw}{Vector of values}

\item{optional:}{data 'All' to return Frequency and Intensity data, Periods to return Intensity only}
}
\value{
periodogram plot (ggplot)
}
\description{
Calculate periodogram
}
