% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{periodogramdata}
\alias{periodogramdata}
\title{Calculate periodogram data}
\usage{
periodogramdata(sbw, smoothing, spli, val = "")
}
\arguments{
\item{sbw}{Vector of values}

\item{smoothing}{window for moving average}

\item{spli}{should be 2}

\item{val}{Title for plot}

\item{averageline}{Line placement for an averageline}
}
\value{
periodogram data
}
\description{
Calculate periodogram data
}
