% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{periodogramlines}
\alias{periodogramlines}
\title{Calculate periodogram}
\usage{
periodogramlines(sbw, smoothing, spli, val = "", averageline = 0)
}
\arguments{
\item{sbw}{Vector of values}

\item{smoothing}{window for moving average}

\item{spli}{should be 2}

\item{val}{Title for plot}

\item{averageline}{Line placement for an averageline}
}
\value{
periodogram plot (ggplot) wilth geom_line()
}
\description{
Calculate periodogram
}
