% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{periodogrammax}
\alias{periodogrammax}
\title{Find peak from periodogram}
\usage{
periodogrammax(sbw, smoothing, spli)
}
\arguments{
\item{sbw}{Vector of values}

\item{smoothing}{window for moving average}

\item{spli}{should be 2}
}
\value{
Frequency of periodogram peak
}
\description{
Find peak from periodogram
}
