% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{periodogramplotdata}
\alias{periodogramplotdata}
\title{Calculate periodogram}
\usage{
periodogramplotdata(
  sbw,
  smoothing,
  spli,
  val = "",
  struc = "list",
  Method = "fft"
)
}
\arguments{
\item{sbw}{Vector of values}

\item{smoothing}{window for moving average}

\item{spli}{should be 2}

\item{val}{Title for plot}

\item{Method}{'fft' or 'Welsh'}

\item{averageline}{Line placement for an averageline}
}
\value{
periodogram data structured for plotting
}
\description{
Calculate periodogram
}
