% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.AOA.R
\name{plot.AOA}
\alias{plot.AOA}
\title{Plot the True IRFs and AOA}
\usage{
plot.AOA(Irf.true, AOA)
}
\arguments{
\item{Irf.true}{a (KxKxh) array generated by get.my.irf() function. The "true" IRFs implied by the DGP parameters.}

\item{AOA}{List generated by AOA.AOA.DGP#}
}
\value{
Creat a ggplot object of class "gg"-"ggplot"
}
\description{
Plot the True IRFs and AOA
}
\examples{
NULL
}
