% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hd.mod.R
\name{plot.hd.mod}
\alias{plot.hd.mod}
\title{plot function for historical decompostion}
\usage{
plot.hd.mod(x)
}
\arguments{
\item{x}{historical decomposition}
}
\value{
plot of historical decomposition
}
\description{
plot function for historical decompostion
}
\examples{
NULL
}
