% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.my.irf.R
\name{plot.my.irf}
\alias{plot.my.irf}
\title{Plot structural impulse response functions}
\usage{
plot.my.irf(
  My.id,
  Step = NULL,
  Partial = NULL,
  Hide.Legend = FALSE,
  Color.manual = NULL,
  Lty.manual = NULL,
  Epsname = NULL,
  Normalize = NULL
)
}
\arguments{
\item{My.id}{Object of class "my.id" or "my.id.boot", identified SVAR}

\item{Step}{Calculation steps}

\item{Partial}{Integer, indexing the partially identified shocks. NULL by default.}

\item{Hide.Legend}{Logical, if Hide.Legend = TRUE, legend is removed}

\item{Color.manual}{Charactuer, set colors manually}

\item{Lty.manual}{Charactuer, set linetype manually}

\item{Epsname}{Characters, Name the structural shocks}

\item{Normalize}{Numeric, normalization of IRFs, IRFs(new) = IRFs / Normalize}
}
\value{
Creat a ggplot object of class "gg"-"ggplot"
}
\description{
Plot structural impulse response functions
}
\examples{
NULL
}
