% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sboot.R
\name{plot.sboot.mod}
\alias{plot.sboot.mod}
\title{modified plot svar.boot}
\usage{
plot.sboot.mod(
  x,
  scales = "free_y",
  lowerq = 0.16,
  upperq = 0.84,
  Partial = NULL,
  Normalize = NULL,
  percentile = "standard",
  ...,
  base
)
}
\arguments{
\item{x}{sboot object}

\item{scales}{NA}

\item{lowerq}{NA}

\item{upperq}{NA}

\item{Partial}{Parial identification}

\item{Normalize}{Numeric, normalization of IRFs, IRFs(new) = IRFs / Normalize}

\item{percentile}{'standard', 'hall' , 'bonferroni'}

\item{base}{NA}
}
\value{
irf plot
}
\description{
modified plot svar.boot
}
\examples{
NA
}
