% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simu.R
\name{plot.simu}
\alias{plot.simu}
\title{Plot identification results of IRFs}
\usage{
plot.simu(IRF.eva)
}
\arguments{
\item{IRF.eva}{A (KxKxh) array generated by functions simulate.**()}
}
\value{
Creat a ggplot object of class "gg"-"ggplot"
}
\description{
Plot identification results of IRFs
}
\examples{
NULL
}
