% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoshenkopelman.R
\name{plottwodPeriodogram}
\alias{plottwodPeriodogram}
\title{Plot periods over time of patch sizes}
\usage{
plottwodPeriodogram(d, type = "image", theta = 320)
}
\arguments{
\item{twodPeriodogram}{output}
}
\value{
Plot
}
\description{
Don't really use anymore
}
