% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.preps.R
\name{preps}
\alias{preps}
\title{Preparation function: check input data, collect meta information, take care of data subsetting. Called by model fitting functions.}
\usage{
preps(env = parent.frame(), call)
}
\description{
Not to be called directly by the user
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
