% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_friday.R
\name{prev_friday}
\alias{prev_friday}
\title{Agrega una columna con el valor encontrado el viernes previo}
\usage{
prev_friday(data, date_col, value_cols, calendar_name = "cal")
}
\arguments{
\item{data}{Dataframe}

\item{date_col}{Nombre de la columna que contiene la fecha}

\item{value_cols}{Columnas para las cuáles se creará una columna con el valor del viernes previo}

\item{calendar_name}{Nombre del calendario a utilizar para ajustar la fecha}
}
\value{
Dataframe con las columnas adicionales
}
\description{
Crea una columna con el valor encontrado el viernes previo (o adía anterior si es feriado) a la fecha. Recibe un dataframe, el nombre de la columna que contiene la fecha, y las columnas para las cuáles creará nuevas columnas, con el sufijo "_lagged". Utiliza la semántica de tidyverse para seleccionar las columnas. Utiliza un calendario para ajustar la fecha en caso de que el viernes sea feriado.
}
