% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{a small functions that calculates and print the progress}
\usage{
progress(x, max = 100, start.time)
}
\arguments{
\item{x}{integer, current iteration step}

\item{max}{integer, total iteration steps}

\item{start.time}{double, starting time}
}
\value{
rest time
}
\description{
a small functions that calculates and print the progress
}
\examples{
NULL
}
