% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUNctions.R
\name{quadInteraction}
\alias{quadInteraction}
\title{Calculating quadratic interactions}
\usage{
quadInteraction(df, binary = FALSE)
}
\arguments{
\item{df}{a data frame containing columns of arbitrary class, however at least one must be numeric}

\item{binary}{do we consider binary columns?}
}
\value{
a data frame containing the interaction terms of the numeric columns of df
}
\description{
This function calculates interactions up to the second order from a data frame
containing at least one numeric column
}
\author{
Philip Goddard
}

