% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi.hd.R
\name{quasi.hd}
\alias{quasi.hd}
\title{quasi historic decomp for a determined period of time}
\usage{
quasi.hd(x, series = 1, Partial = NULL, Epsname = NULL, Start, End, Freq = 4)
}
\arguments{
\item{x}{svar or my.id}

\item{series}{which series to be decomposed}

\item{Partial}{which shock}

\item{Epsname}{shock name}

\item{Start}{Start of period}

\item{End}{End of period}

\item{Freq}{Frequency of data}
}
\description{
quasi historic decomp for a determined period of time
}
\examples{
NULL
}
