% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUNctions.R
\name{rangebin}
\alias{rangebin}
\title{Bin into ranges. Simple wrapper around cut3}
\usage{
rangebin(input, bins, allowNK = FALSE)
}
\arguments{
\item{input}{numeric input}

\item{bins}{vector specifying the breaks}

\item{allowNK}{logical. If TRUE, a level for NK (not known) will be generated from NA's}
}
\description{
Bin into ranges. Simple wrapper around cut3
}

