% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_ac.R
\name{Remove_AC}
\alias{Remove_AC}
\title{Remove Autocorrelation in IV}
\usage{
Remove_AC(iv, lag)
}
\arguments{
\item{iv}{Instruments, ts object}

\item{lag}{}
}
\value{

}
\description{
Remove Autocorrelation in IV
}
\examples{
NULL
}
