% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmat.manipulation.R
\name{Reorder.Frob}
\alias{Reorder.Frob}
\title{Reordering columns by minimizing the sqaured Frob. distance}
\usage{
Reorder.Frob(B.bench, B.nonBench)
}
\arguments{
\item{B.bench}{KxK benchmark decomposition Matrix}

\item{B.nonBench}{KxK decomposition Matrix which this function is applied to}
}
\value{
reordered KxK benchmark decomposition Matrix
}
\description{
Reordering columns by minimizing the sqaured Frob. distance
}
\examples{
NULL
}
