% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleDF.R
\name{rescaleDF}
\alias{rescaleDF}
\title{RescaleDF}
\usage{
rescaleDF(df, date, column, format)
}
\arguments{
\item{df}{El dataframe que contiene los datos}

\item{date}{La fecha a buscar y que será el pivot}

\item{column}{en qué columna está el valor que será seteado a 100}

\item{format}{0 Si es wide o la columna donde están los nombres en caso que sea long format}
}
\value{
devuelve el df reescalado
}
\description{
Reescala una serie de tiempo en función de un valor de la serie.
Hacia adelante y hacia atrás.
También ajusta las series contiguas si las hay.
Es ideal para cuando hay series de tiempo juntas en un df, se le
pasa el df a la función y reescalará todo en función de una fecha.
}
