% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bmat.manipulation.R
\name{resign.mat}
\alias{resign.mat}
\title{Resign a matrix such that diagonal elements are all positive}
\usage{
resign.mat(Bmat, K)
}
\arguments{
\item{Bmat}{Matrix, a (KxK) input matix}

\item{K}{Integer, dimension of matrix}
}
\value{
Matrix, a (KxK) output matix
}
\description{
Resign a matrix such that diagonal elements are all positive
}
\examples{
NULL
}
