% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_gene.R
\name{resolve_gene}
\alias{resolve_gene}
\title{Verify a gene matches our reference syntax}
\usage{
resolve_gene(input_gene, chain)
}
\arguments{
\item{input_gene}{A character vector.}

\item{chain}{"trav", "trbv", "traj", or "trbj"}
}
\value{
A character vector in uppercase.
}
\description{
Verify a gene matches our reference syntax
}
