% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatioTemporalCorrelation.R
\name{rij.corr}
\alias{rij.corr}
\title{Calculate spatiotemporal cross correlation}
\usage{
rij.corr(
  Array,
  ms,
  N,
  dclasses,
  parallel = T,
  ncores = 4,
  verbose = F,
  timelag = 0,
  SecondArray = NA
)
}
\arguments{
\item{Array}{3-D array of landscape snapshots (from OutbreakLandscape Files)}

\item{ms}{Number of smaples to take from the landscape}

\item{N}{Number of points in each sample}

\item{parallel}{Boolean. Calculate in parallel? Default=T}

\item{ncores}{Number of cores available for parallel computing. Default=4}

\item{verbose}{Print steps}

\item{timelag}{Increase timelag. Default=0}

\item{SecondArray}{3-D array of landscape snapshots for another species (from OutbreakLandscape Files)}

\item{dclassess}{Distance classes to investigate}
}
\value{
Data frame for correlogram of Rij
}
\description{
Calculate spatiotemporal cross correlation
}
