% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sd.R
\name{sample_sd}
\alias{sample_sd}
\title{Calculate the sample standard deviation}
\usage{
sample_sd(x)
}
\arguments{
\item{x}{a vector of numbers of length N}
}
\value{
the sample sd of x
}
\description{
Calculate the sample standard deviation
}
\details{
This function calculate the sample standard deviation
}
\examples{
a <- c(1,2,3,4,5)
sample_sd(a)

}
