% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seasons_Stats_NBA.R
\docType{data}
\name{Seasons_Stats_NBA}
\alias{Seasons_Stats_NBA}
\title{The data-set contains aggregate individual statistics for 67 NBA seasons.}
\format{
A tibble with 24691 obs and 52 variables:
\describe{
\item{}{}
\item{}{}
...
}
}
\source{
\url{https://www.kaggle.com/datasets/drgilermo/nba-players-stats}
}
\usage{
Seasons_Stats_NBA
}
\description{
from basic box-score attributes such as points, assists, rebounds etc.,
to more advanced money-ball like features such as Value Over Replacement.
}
\keyword{datasets}
