% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.wild.fixed.Boots.R
\name{signcheck.sboot}
\alias{signcheck.sboot}
\title{Check sboot sign}
\usage{
signcheck.sboot(x)
}
\arguments{
\item{x}{object "sboot"}
}
\value{

}
\description{
Check sboot sign
}
\examples{
NULL
}
