% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.datas.R
\name{simu.datas}
\alias{simu.datas}
\title{Simulate Data from given DGP}
\usage{
simu.datas(Config, DGP, ...)
}
\arguments{
\item{Config}{List, configuration for simulation}

\item{DGP}{Integer, indicates DGP}

\item{...}{Other parameters for DGP}
}
\value{
A list containing simulated dats.
}
\description{
Simulate Data from given DGP
}
\examples{
NULL
}
