% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.chol.R
\name{simulate.chol}
\alias{simulate.chol}
\title{Recursive identification and evaluation based on Cholesky decomposition}
\usage{
\method{simulate}{chol}(
  Data,
  AOA,
  Subsample = NULL,
  Plot = FALSE,
  Varname = c("x", "pi", "r")
)
}
\arguments{
\item{Data}{List containing simulated data}

\item{AOA}{List generated by function AOA.3d()}

\item{Subsample}{Vector, a (1x2) vector containing start- and end-point of subsamples that will be
subject to estimation and identification}

\item{Plot}{Logical, if Plot = TRUE, a plot of identification performance will be created}

\item{Varname}{Character, names of variables in the systems}
}
\value{
A (KxKxh) array containing proportion of all simulated IRFs that lie in the AOA
}
\description{
Recursive identification and evaluation based on Cholesky decomposition
}
\examples{
NULL
}
