% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.cv.R
\name{simulate.cv}
\alias{simulate.cv}
\title{Identification and evaluation based on heteroscedasticity (shift in variance)}
\usage{
\method{simulate}{cv}(
  Data,
  AOA,
  Reorder = TRUE,
  Subsample = NULL,
  Break,
  Guess = FALSE,
  Bmat,
  Plot = FALSE,
  Varname = c("x", "pi", "r")
)
}
\arguments{
\item{Data}{List containing simulated data}

\item{AOA}{List generated by function AOA.3d()}

\item{Reorder}{Logical, if Reorder = TRUE, columns of B matrix are reorded according to AOA}

\item{Subsample}{Vector, a (1x2) vector containing start- and end-point of subsamples that will be
subject to estimation and identification}

\item{Break}{Numeric, integer or a (2x1) vector indicates the time point(s) of structural break(s)}

\item{Guess}{Logical, if Guess = True, the econometrician will make a consistent guess about the time point of structural break}

\item{Bmat}{Matrix of dimension (KxK), the true structural decomposition implied by the DGP}

\item{Plot}{Logical, if Plot = TRUE, a plot of identification performance will be created}

\item{Varname}{Character, names of variables in the systems}
}
\value{
A (KxKxh) array containing proportion of all simulated IRFs that lie in the AOA
}
\description{
Identification and evaluation based on heteroscedasticity (shift in variance)
}
\examples{
NULL
}
