% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.dcov.R
\name{simulate.dcov}
\alias{simulate.dcov}
\title{Independence-based identification and evaluation based on distance covariances}
\usage{
\method{simulate}{dcov}(
  Data,
  AOA,
  Reorder = TRUE,
  Bmat,
  Subsample = NULL,
  Plot = FALSE,
  Varname = c("x", "pi", "r")
)
}
\arguments{
\item{Data}{List containing simulated data}

\item{AOA}{List generated by function AOA.3d()}

\item{Reorder}{Logical, if Reorder = TRUE, columns of B matrix are reorded according to AOA}

\item{Bmat}{Matrix of dimension (KxK), the true structural decomposition implied by the DGP}

\item{Subsample}{Vector, a (1x2) vector containing start- and end-point of subsamples that will be
subject to estimation and identification}

\item{Plot}{Logical, if Plot = TRUE, a plot of identification performance will be created}

\item{Varname}{Character, names of variables in the systems}
}
\value{
A (KxKxh) array containing proportion of all simulated IRFs that lie in the AOA
}
\description{
Independence-based identification and evaluation based on distance covariances
}
\examples{
NULL
}
