% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.iv.R
\name{simulate.iv}
\alias{simulate.iv}
\title{Identification and evaluation based on Proxy SVAR with given instruments}
\usage{
\method{simulate}{iv}(
  Data,
  AOA,
  Weak = FALSE,
  Combi = FALSE,
  Bmat,
  Subsample = NULL,
  Plot = FALSE,
  Varname = c("x", "pi", "r")
)
}
\arguments{
\item{Data}{List containing simulated data}

\item{AOA}{List generated by function AOA.3d()}

\item{Weak}{Logical, if Weak = TRUE, weak instruments will be used for identification}

\item{Combi}{Logical, if Combi = TRUE, instruments that are correlated with more than one endogeneous
structural shocks will be used for identification}

\item{Bmat}{Matrix of dimension (KxK), the true structural decomposition implied by the DGP}

\item{Subsample}{Vector, a (1x2) vector containing start- and end-point of subsamples that will be
subject to estimation and identification}

\item{Plot}{Logical, if Plot = TRUE, a plot of identification performance will be created}

\item{Varname}{Character, names of variables in the systems}
}
\value{
A (KxKxh) array containing proportion of all simulated IRFs that lie in the AOA
}
\description{
Identification and evaluation based on Proxy SVAR with given instruments
}
\examples{
NULL
}
