% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.sign.R
\name{simulate.sign}
\alias{simulate.sign}
\title{Identification and evaluation based on Sign restrictions}
\usage{
\method{simulate}{sign}(
  Data,
  AOA,
  Subsample = NULL,
  Iter,
  Itermax,
  Signmat,
  TrueSign,
  RHorizont,
  MT = FALSE,
  PositiveDiag = TRUE,
  Ci,
  Varname = c("x", "pi", "r"),
  Bmat,
  Plot = FALSE,
  Hist = FALSE
)
}
\arguments{
\item{Data}{List containing simulated data}

\item{AOA}{List generated by function AOA.3d()}

\item{Iter}{Integer, number of rotations}

\item{Itermax}{Integer, number of maximum draws}

\item{Signmat}{Matrix, a (KxK) matrix containing Sign restrictions. 1: positive sign; -1: negative sign; NA: unrestricted}

\item{TrueSign}{Matrix, a (KxK) matrix containing the true Sign pattern in DGP 1: positive sign; -1: negative sign}

\item{RHorizont}{Integer, RHorizonts of impulse responses, for which the sign restrictions are imposed}

\item{MT}{Logical, if MT = TRUE, Fry and Pagan’s (2011) median target (MT) method will be used}

\item{PositiveDiag}{Logical, if PositiveDiag = TRUE, the diagonal of impact matrix will be forced to be positive}

\item{Ci}{Double, between [0,1] sample quantiles for IRFs}

\item{Varname}{Character, names of variables in the systems}

\item{Bmat}{Matrix of dimension (KxK), the true structural decomposition implied by the DGP}

\item{Plot}{Logical, if Plot = TRUE, a plot of identification performance will be created#' @param Subsample Vector, a (1x2) vector containing start- and end-point of subsamples that will be
subject to estimation and identification}

\item{Hist}{Logical, if Hist = TRUE, histogram of distribution of freqencies of detecting the correct sign for the unrestricted elements in the identified B matrices is provided}
}
\value{
A (KxKxh) array containing proportion of all simulated IRFs that lie in the AOA
}
\description{
Identification and evaluation based on Sign restrictions
}
\examples{
NULL
}
