% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoshenkopelman.R
\name{sizes}
\alias{sizes}
\title{Returns bins of sizes of patches.}
\usage{
sizes(d)
}
\arguments{
\item{d}{Output from hosen()}
}
\value{
Binned sizes of patches
}
\description{
Returns bins of sizes of patches.
}
