% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoshenkopelman.R
\name{SpatialHistogram}
\alias{SpatialHistogram}
\title{Histogram of patch sizes at one time step}
\usage{
SpatialHistogram(d, year, ret = "plot")
}
\arguments{
\item{d}{OutbreakLandscape file}

\item{year}{row to generate histogram for}

\item{ret}{Return "plot" or "matrix"}
}
\value{
Histogram of patch sizes
}
\description{
Histogram of patch sizes at one time step
}
