% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoshenkopelman.R
\name{SpatialHistogramOnSquare}
\alias{SpatialHistogramOnSquare}
\title{Histogram of patch sizes at one time step from already squared data}
\usage{
SpatialHistogramOnSquare(d, ret = "plot")
}
\arguments{
\item{d}{Landscape snapshot}
}
\value{
Histogram of patch sizes
}
\description{
Histogram of patch sizes at one time step from already squared data
}
