% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoshenkopelman.R
\name{square}
\alias{square}
\title{Makes a line of data from OutbreakLandscape files into a landscape snapshot matrix}
\usage{
square(d, rows = 0, cols = 0)
}
\arguments{
\item{d}{Single line from OutbreakLandscape file}
}
\value{
Matrix of snapshot in tim from the model
}
\description{
Makes a line of data from OutbreakLandscape files into a landscape snapshot matrix
}
