% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{stepwise}
\alias{stepwise}
\title{Stepwise selection for AIC/BIC (Wrapper for MASS package stepAIC())}
\usage{
stepwise(x, direction = "backwards", test)
}
\arguments{
\item{x}{Model object}

\item{direction}{Direction for searching}

\item{test}{AIC or BIC}
}
\value{
AICModel
}
\description{
Stepwise selection for AIC/BIC (Wrapper for MASS package stepAIC())
}
