% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumarisa_periodico.R
\name{sweek}
\alias{sweek}
\title{Sumariza data por semana}
\usage{
sweek(df, date_column = "date", value_columns = everything())
}
\arguments{
\item{df}{Un dataframe}

\item{date_column}{El nombre de la columna que contiene la fecha. "date" por default}

\item{value_columns}{Las columnas que se desean sumarizar. Todas las columnas por default}
}
\value{
Un dataframe con la sumarización de los datos por semana.
}
\description{
Esta functión recibe un dataframe y sumariza los datos por semana, calculando la suma, media, mínimo y mnáximo.
}
\examples{
sweek(ccl, date_column = "date", value_columns = c("mepAL", "ccl3"))
}
