% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate.simu.aoa}
\alias{tabulate.simu.aoa}
\title{Show average coverage frequencies of AoAs as table}
\usage{
tabulate.simu.aoa(Simus, s = 4, l = 15, Label, Varname = c("x", "pi", "r"))
}
\arguments{
\item{Simus}{List, contains multiple (KxKxh) arrays generated by functions simulate.**()}

\item{s}{Numeric, short-run period up to...}

\item{l}{Numeric, long-run period up to...}

\item{Label}{Character, names of identification methods}

\item{Varname}{Character, names of variables in the systems}
}
\value{
List containing average coverage frequencies discriminate btw short- and long-run
}
\description{
Show average coverage frequencies of AoAs as table
}
\examples{
NULL
}
