% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate.simu.ump}
\alias{tabulate.simu.ump}
\title{Show frequencies of identification of UMP and RMSE as table}
\usage{
tabulate.simu.ump(Simus, Label, Varname = c("x", "pi", "r"))
}
\arguments{
\item{Simus}{List, contains multiple (KxKxh) arrays generated by functions simulate.**()}

\item{Label}{Character, names of identification methods}

\item{Varname}{Character, names of variables in the systems}
}
\value{
List
}
\description{
Show frequencies of identification of UMP and RMSE as table
}
\examples{
NULL
}
