% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.GC.R
\name{test.GC}
\alias{test.GC}
\title{Test for Granger Causality}
\usage{
test.GC(x, C1)
}
\arguments{
\item{x}{Class of "varest"}

\item{C1}{a KxK matrix contains the restrictions}
}
\value{
test statistics and p value
}
\description{
Test for Granger Causality
}
\examples{
NULL
}
