% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.indep.R
\name{test.indep}
\alias{test.indep}
\title{Calculate dcov and cvm distance of choleski and optimized decompositions}
\usage{
test.indep(u, Bmat, dd = NULL)
}
\arguments{
\item{u}{T by K data matrix.}

\item{Bmat}{K by K optimized decomposition matrix.}

\item{dd}{Object of class 'indepTestDist' (generated by 'indepTest' from package 'copula').}
}
\value{
dcov and cvm distance of choleski and optimized decompositions
}
\description{
Calculate dcov and cvm distance of choleski and optimized decompositions
}
\examples{
NULL
}
