% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.weak.IV.R
\name{test.weak.IV}
\alias{test.weak.IV}
\title{Test for weak instruments in Lunsford 2016}
\usage{
test.weak.IV(Tob, u, instruments)
}
\arguments{
\item{Tob}{Numeric, length of simulated time series}

\item{u}{Vector, a (Tx1) vector containing residuals from reduced form VAR}

\item{instruments}{Vector of dimension (Tx1) containing SINGLE instrument for ONE of structural shocks
with the same length as the time series}
}
\value{
Numeric, test statistic. For critical value: look at Lunsford 2016: p.16 Table 2
}
\description{
Test for weak instruments in Lunsford 2016
}
\examples{
NULL
}
