% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoshenkopelman.R
\name{twodPeriodogram}
\alias{twodPeriodogram}
\title{Calculate periods over time of patch sizes}
\usage{
twodPeriodogram(d)
}
\arguments{
\item{OutbreakLandscape}{file}
}
\value{
Periodogram data
}
\description{
Calculate periods over time of patch sizes
}
