% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{univariatecox}
\alias{univariatecox}
\title{Univariate cox model for apply function}
\usage{
univariatecox(x)
}
\arguments{
\item{x}{Cox model}
}
\value{
Formatted Cox Model
}
\description{
Univariate cox model for apply function
}
