% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wavelengthtoindex}
\alias{wavelengthtoindex}
\title{\code{wavelengthtoindex} converts wavelength (nm) into an index number for the terraref dataset.}
\usage{
wavelengthtoindex(wavelength)
}
\arguments{
\item{wavelength}{A wavelenghth nm}
}
\value{
an index
}
\description{
\code{wavelengthtoindex} converts wavelength (nm) into an index number for the terraref dataset.
}
\examples{
wavelengthtoindex(450)
}
