% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wia.R
\name{wia}
\alias{wia}
\title{Willmot Index of Agreement (d)}
\usage{
wia(obs, sim)
}
\arguments{
\item{obs}{observed data vector}

\item{sim}{simulated data vector}
}
\value{
(d) the index of agreement
}
\description{
wia calculates the Willmot index of agreement of simulated to observed data.
}
