% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{AveragePeriodogram}
\alias{AveragePeriodogram}
\title{Calculate Average periodogram over whole landscape from aggregate data in 'Output files'}
\usage{
AveragePeriodogram(
  DataDir,
  Variable,
  Value,
  Species,
  globr,
  reps,
  burnin,
  plotmax = 0,
  Names = 0,
  Title = T,
  yl = T,
  data = F,
  Density = F,
  Method = "fft",
  ggpltstring = ""
)
}
\arguments{
\item{DataDir}{Directory containing the data}

\item{Variable}{Name of the Variable in the filename}

\item{Species}{Species to use. Default=SBW}

\item{globr}{Number of Simulations to average}

\item{reps}{Length of simulations}

\item{burnin}{Burnin period}

\item{plotmax}{optional: Ymax height of plot}

\item{Names}{sometimes optional: Vection of names from data file}

\item{Title}{optional: Plot title}

\item{yl}{Remove y label}

\item{data}{\itemize{
\item optional: Return plot data rather than grob
}}

\item{Density}{\itemize{
\item optional: Value of total area for density calculations
}}

\item{Method}{\itemize{
\item method for transformation. Default = 'fft'. Options: 'fft', 'Walsh'
}}

\item{Values}{Vector of parameter values for the Moran's I to be calculated for}
}
\value{
Grob with the average peiodogram plotted if data = F, data.frame containing the data used to produce grob if data = T
}
\description{
Calculate Average periodogram over whole landscape from aggregate data in 'Output files'
}
