% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDLR.R
\name{CCLGGAL}
\alias{CCLGGAL}
\title{CCLGGAL}
\usage{
CCLGGAL(from = Sys.Date(), to = Sys.Date() + 1)
}
\arguments{
\item{from}{Fecha desde}

\item{to}{Fecha hasta}
}
\value{
Devuelve un tibble con la serie en cuestión
}
\description{
Devuelve el CCL calculado con GGAL.
}
\examples{
CCLGGAL(from = "2020-09-14", to = sys.Date() + 1) devuelve el CCL via GGAL
}
