% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clustersub}
\alias{clustersub}
\title{\code{clustersub} Function to use data from kmeans cluster to subset data}
\usage{
clustersub(x, fit, cluster = 1)
}
\arguments{
\item{x}{a vector from a raster}

\item{fit}{a kmeans model}

\item{cluster}{a specific kemans cluster}
}
\description{
\code{clustersub} Function to use data from kmeans cluster to subset data
}
