% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbConnectP}
\alias{dbConnectP}
\title{dbConnectP}
\usage{
dbConnectP(host = "10.192.97.146", port = 5432, dbname = "data")
}
\description{
Wrapper de dbConnect pero con los parámetros de servidor, puerto, user y password ya establecidos.
host <- "10.192.97.146"
port <- 5432
dbname <- 'data'
user <- Sys.getenv("POSTGRES_USER")
password <- Sys.getenv("POSTGRES_PASS")
}
\examples{
con <- dbConnectP()

}
