% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbExecuteQuery}
\alias{dbExecuteQuery}
\title{executeQuery}
\usage{
dbExecuteQuery(
  query,
  host = Sys.getenv("POSTGRES_HOST"),
  port = 5432,
  dbname = Sys.getenv("POSTGRES_DB"),
  user = Sys.getenv("POSTGRES_USER"),
  password = Sys.getenv("POSTGRES_PASSWORD")
)
}
\arguments{
\item{query}{String con la query. Atención que tablas van con double quotes y valores con single. Recomiendo usar
single quotes afuera y double adentro.}

\item{host}{IP del servidor postgreSQL}

\item{port}{Puerto del servidor}

\item{dbname}{Base de Datos. Default='data'}

\item{user}{Usuario de la base de datos}

\item{password}{Password de la base de datos en el servidor}
}
\description{
Es un mero wraper de dbGetQuery de DBI para no tener que indicar servidor, puerto y db
}
\examples{
executeQuery(query = 'SELECT * FROM "A3500"')


}
