% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbExecuteQuery}
\alias{dbExecuteQuery}
\title{executeQuery}
\usage{
dbExecuteQuery(query, host = "10.192.97.146", port = 5432, dbname = "data")
}
\arguments{
\item{query}{String con la query. Atención que tablas van con double quotes y valores con single. Recomiendo usar
single quotes afuera y double adentro.}

\item{host}{IP del servidor postgresQL}

\item{port}{Puerto del servidor}

\item{dbname}{Base de Datos. Default='data'}
}
\description{
Es un mero wraper de dbGetQuery de DBI para no tener que indicar servidor, puerto y db
}
\examples{
executeQuery(query = 'SELECT * FROM "A3500"')


}
