% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbExecuteQuery}
\alias{dbExecuteQuery}
\title{dbExecuteQuery}
\usage{
dbExecuteQuery(
  query,
  server = "local",
  host = NULL,
  port = NULL,
  db = NULL,
  user = NULL,
  password = NULL
)
}
\arguments{
\item{query}{String con la query. Atención que tablas van con double quotes y valores con single. Recomiendo usar
single quotes afuera y double adentro.}

\item{host}{IP del servidor postgreSQL}

\item{port}{Puerto del servidor}

\item{db}{Base de Datos. Default='data'}

\item{user}{Usuario de la base de datos}

\item{password}{Password de la base de datos en el servidor}
}
\value{
\code{data.frame} con el resultado de la consulta.
}
\description{
Es un mero wraper de dbGetQuery de DBI para no tener que indicar servidor, puerto y db
}
\examples{
\dontrun{dbExecuteQuery(query = 'SELECT * FROM "A3500"')}

}
\seealso{
getConf
}
