% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbGetTable}
\alias{dbGetTable}
\title{getTableP}
\usage{
dbGetTable(
  table,
  host = Sys.getenv("POSTGRES_HOST"),
  port = 5432,
  dbname = Sys.getenv("POSTGRES_DB"),
  user = Sys.getenv("POSTGRES_USER"),
  password = Sys.getenv("POSTGRES_PASSWORD")
)
}
\arguments{
\item{table}{Acá hay que poner alguna. Controla que no esté vacío}

\item{host}{IP del servidor postgreSQL}

\item{port}{Puerto del servidor}

\item{dbname}{Base de Datos. Default='data'}

\item{user}{Usuario de la base de datos}

\item{password}{Password de la base de datos en el servidor}
}
\description{
Función para traer una tabla desde el servidor postgresQL
}
