% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbGetTable}
\alias{dbGetTable}
\title{dbGetTable}
\usage{
dbGetTable(
  table,
  server = "local",
  host = NULL,
  port = NULL,
  dbname = NULL,
  user = NULL,
  password = NULL
)
}
\arguments{
\item{table}{Acá hay que poner alguna. Controla que no esté vacío}

\item{server}{permite evitar los demás parámetros del servidor y que los obtenga con getConf}

\item{host}{IP del servidor postgreSQL}

\item{port}{Puerto del servidor}

\item{dbname}{Base de Datos. Default='data'}

\item{user}{Usuario de la base de datos}

\item{password}{Password de la base de datos en el servidor}
}
\description{
Función para traer una tabla desde el servidor postgresQL
}
