% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbWriteDF}
\alias{dbWriteDF}
\title{dbWriteDF}
\usage{
dbWriteDF(
  table,
  df,
  host = Sys.getenv("POSTGRES_HOST"),
  port = 5432,
  dbname = Sys.getenv("POSTGRES_DB"),
  user = Sys.getenv("POSTGRES_USER"),
  password = Sys.getenv("POSTGRES_PASSWORD")
)
}
\arguments{
\item{host}{IP del servidor postgreSQL}

\item{port}{Puerto del servidor}

\item{dbname}{Base de Datos. Default='data'}

\item{user}{Usuario de la base de datos}

\item{password}{Password de la base de datos en el servidor}
}
\description{
Wraper de dbWriteTable del paquete DBI pero que asume las conexiones de mi servidor.
}
\examples{
dbWriteDF(table = 'data', df = dfAGrabar)
}
