% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variasPostgres.R
\name{dbWriteDF}
\alias{dbWriteDF}
\title{dbWriteDF}
\usage{
dbWriteDF(
  table,
  df,
  server = "local",
  host = NULL,
  port = NULL,
  db = NULL,
  user = NULL,
  password = NULL,
  ...
)
}
\arguments{
\item{server}{permite evitar los demás parámetros del servidor y que los obtenga con getConf}

\item{host}{IP del servidor postgreSQL}

\item{port}{Puerto del servidor}

\item{user}{Usuario de la base de datos}

\item{password}{Password de la base de datos en el servidor}

\item{dbname}{Base de Datos. Default='data'}
}
\description{
Wraper de dbWriteTable del paquete DBI pero que asume las conexiones de mi servidor.
}
\examples{
\dontrun{dbWriteDF(table = 'data', df = dfAGrabar)}
}
