% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos.R
\name{fn_cos}
\alias{fn_cos}
\title{Compute the cosine}
\usage{
fn_cos(x, k)
}
\arguments{
\item{x}{numeric, the number to be transformed}

\item{k}{integer, the number of terms to be used in the series expansion beyond the constant 1. The value of k is always >= 1.}
}
\value{
the cosine value of input number
}
\description{
Compute the cosine of a number using the truncated series expansion
}
\details{
This function compute the cosine of a number
}
\examples{
fn_cos(1,5)

}
