% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMerval.R
\name{getMerval}
\alias{getMerval}
\title{getMerval devuelve el Merval y otros valores}
\usage{
getMerval(fechaInicio = "2014-05-27", ...)
}
\description{
getMerval devuelve el Merval y otros valores
}
\details{
Trae merval en pesos, ccl (via api rofex más histórico), USCPI mervalCCL y mervalCCLAjustado

@param FechaInicio Fecha por defecto 2014-05-27
@returns Una tibble de 6 columnas

@examples \dontrun{getMerval()}
}
