% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\title{Trae una tabla de las bases de datos locales}
\usage{
getTable(table, file = "~/data/test.sqlite3")
}
\description{
Trae una tabla de las bases de datos locales
}
\details{
Simplifica el trae una tabla de las bases locales.
Cada tabla tiene una db asociada. En la base test
La función busca la tabla y abre la DB que corresponda

 @param Tabla No tiene valor por defecto
 @param File permite cambiar el path y nombre de la db a buscar
 Sino va al default que se llama "test"
 @returns Una tibble igual a la tabla consultada

 @examples getTable("tabla-a-buscar") -> Devuelve la tibble
}
