% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\title{Trae una tabla de las bases de datos locales}
\usage{
getTable(table = NULL, overrideDates = FALSE, file = "~/data/test.sqlite3")
}
\arguments{
\item{table}{No tiene valor por defecto}

\item{overrideDates}{Si TRUE, no convierte los campos Fecha}

\item{file}{permite cambiar el path y nombre de la db a buscar. Si no, usa "test"}
}
\value{
Una tibble igual a la tabla consultada
}
\description{
Trae una tabla de las bases de datos locales
}
\details{
Simplifica el trae una tabla de las bases locales.
Cada tabla tiene una db asociada. En la base test
La función busca la tabla y abre la DB que corresponda.
}
\examples{
\dontrun{ getTable("tabla-a-buscar") }
}
