% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USCPDataI.R
\name{getUSCPI}
\alias{getUSCPI}
\title{getUSCPI}
\usage{
getUSCPI(format = "", db = "", ...)
}
\arguments{
\item{format}{"daily" or "monthly". Default is "" which returns the raw data.}

\item{db}{"" for Postgres or "sqlite" for local sqlite database.}
}
\value{
tibble with the CPI data.
}
\description{
Wrapper to get the US CPI data from local database. It can return daily or monthly data.
}
