% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYields.R
\name{getYields}
\alias{getYields}
\title{getYieds}
\usage{
getYields(
  letras,
  settlementDate,
  precios,
  initialFee = 0,
  endingFee = 0,
  endpoint = "yield"
)
}
\arguments{
\item{letras}{Puede ser un vector conteniendo múltiples}

\item{settlementDate}{Fecha en formato "yyyy-mm-dd"}

\item{precios}{Puede ser vector.}

\item{initialFee}{Valor único a aplicar upfront. Se recicla}

\item{endingFee}{Valor único a aplicar al final (si se intenta estimar una venta posterio)}

\item{endpoint}{"yield" o "apr". Dependiendo a qué endpoint del servidor quiera pegársele.}
}
\value{
tibble with same data plus yield & duration. There are several other values that the API
returns but not implemented yet.
}
\description{
getYieds
}
\details{
Function to get yields from the local server running yields.go at 127.0.0.1:8080.
Fees are recycled for all the bonds in the array.
}
\examples{
getYields("GD30D", "2023-07-13", 32.3, initialFee = 0.007515, endingFee = 0, endpoint = "yield")

}
