% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabaGrafo.R
\name{grabaGrafo}
\alias{grabaGrafo}
\title{Guarda un gráfico en un archivo}
\usage{
grabaGrafo(
  fecha = Sys.Date(),
  variable,
  path = "~/OneDrive/outlier/docs/tablasGraficos",
  width = 14.2,
  height = 7.61,
  dpi = 600,
  units = "in"
)
}
\description{
Esta función guarda un gráfico en un archivo con un nombre que incluye la fecha y el nombre de la variable.
}
\details{
@param fecha Fecha que se incluirá en el nombre del archivo. Por defecto es la fecha actual.
 @param variable Gráfico que se guardará.
 @param path Ruta donde se guardará el archivo. Por defecto es "~/OneDrive/outlier/docs/tablasGraficos".
 @param width Ancho del gráfico en pulgadas. Por defecto es 14.2.
 @param height Alto del gráfico en pulgadas. Por defecto es 7.61.
 @param dpi Resolución del gráfico. Por defecto es 300.
 @param units Unidades del gráfico. Por defecto es "in".
 @return NULL
 @export
 @examples
 \dontrun{
 library(ggplot2)
 library(outlier)
 data("datos")
 ggplot(datos, aes(x = fecha, y = valor)) + geom_line()
 grabaGrafo(variable = ggplot(datos, aes(x = fecha, y = valor)) + geom_line())
 }
 @import ggplot2
 @importFrom utils deparse substitute
 @importFrom base format
 @importFrom base as.Date
 @importFrom ggplot2 ggsave
 @importFrom base paste0
 @importFrom base file.path
 @importFrom base missing
}
