% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_icon.R
\name{gradient_icon}
\alias{gradient_icon}
\title{gradient_icon}
\usage{
gradient_icon(col = colorRampPalette(c("grey90", "grey20"))(100),
  labels = c(0, 1), tdist = 0.2, bars = 10, title = NULL,
  dist = c(0.4, 0.2), tcol = "white", filename = NULL, ratio = 0.5)
}
\arguments{
\item{col}{spatialpointsdataframe with the parameters}

\item{labels}{minimum and maximum value}

\item{tdist}{distance of the labels from the gradient scale}

\item{bars}{number of lines in the gradient scale}

\item{title}{title of the gradient scale}

\item{dist}{distance from x and y top corner}

\item{tcol}{color of labels and title}

\item{filename}{the name of the legend_icon when saving as a png, if NULL the png is not exported}

\item{ratio}{height-width ratio}
}
\description{
Generates a gradient scale
}
\examples{
gradient_icon(col=colorRampPalette(c('grey90','grey20'))(100), labels = c(500,1000), tdist=0.2,bars=10,title=NULL,filename='gradient_icon',res2=30,ratio=0.5) 
}
\keyword{gradient,}
\keyword{plot}
