% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{icon}
\alias{icon}
\title{icon}
\usage{
icon(pch = 24, col = "black", bg = "white")
}
\arguments{
\item{pch}{symbol to use}

\item{col}{color border}

\item{bg}{color background}
}
\description{
generate an icon
}
\examples{
# generate and export an icon
export(func=icon(pch=21, col='black',bg='white'),'icon',type='png') 
}
\keyword{export,}
\keyword{icon,}
\keyword{png}
