% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeimage.R
\name{makeimage}
\alias{makeimage}
\title{Make raster plot of an OutbreakLandscape file}
\usage{
makeimage(d, V, v, s = 200, p = "SBW")
}
\arguments{
\item{d}{Directory containing data}

\item{V}{Variable of interest}

\item{v}{value of interest}

\item{s}{landscape size. Default=200}

\item{p}{Species. Default=SBW}
}
\value{
Histogram of patch sizes
}
\description{
Make raster plot of an OutbreakLandscape file
}
