% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeimage.R
\name{makeimage}
\alias{makeimage}
\title{Make raster plot of an OutbreakLandscape file}
\usage{
makeimage(d, V, v, s = 200, p = "SBW", t = F, timestep = 50, log = T)
}
\arguments{
\item{d}{Directory containing data.}

\item{V}{Variable of interest.}

\item{v}{vector of values.}

\item{s}{landscape size. Default=200.}

\item{p}{Species. Default=SBW.}

\item{t}{Optional. Title for the plot. Default: False.}

\item{timestep}{Optional. Timestep of the data file to plot. Default: 50}

\item{log}{Optional. Log the values before plotting.}
}
\value{
Raster image of OutbreakLandscape file.
}
\description{
Make raster plot of an OutbreakLandscape file
}
