% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalMeans.R
\name{Means}
\alias{Means}
\title{Summary statistics for global temporal data (Output...)}
\usage{
Means(workingdir, vari, i, r, fun = "mean")
}
\arguments{
\item{workingdir}{Directory the data is in}

\item{vari}{Variable}

\item{i}{Value}

\item{r}{Total number of simulations}

\item{fun}{Function}
}
\value{
The number of slices in the array
}
\description{
Returns summary statistics for outputs of the model from Output... files
}
