% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Calculate periodogram}
\usage{
periodogram(sbw, smoothing, spli = 2, val = "", averageline = 0)
}
\arguments{
\item{sbw}{Vector of values}

\item{smoothing}{window for moving average}

\item{spli}{should be 2}

\item{val}{Title for plot}

\item{averageline}{Line placement for an averageline}
}
\value{
periodogram plot (ggplot)
}
\description{
Calculate periodogram
}
