% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hd.mod.R
\name{plot.hd.mod}
\alias{plot.hd.mod}
\title{Title}
\usage{
plot.hd.mod(x)
}
\arguments{
\item{x}{historical decomposition}
}
\value{
plot of historical decomposition
}
\description{
Title
}
\examples{
NULL
}
