% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\title{Residual plots}
\usage{
plot_lm(x, plots = c(1L:4L), id.n = 3, labels.id = names(residuals(x)),
  caption = NULL, arrange = T, ...)
}
\arguments{
\item{x}{an lm model object}

\item{plots}{}

\item{id.n}{}

\item{labels.id}{}

\item{caption}{}

\item{arrange}{}

\item{...}{}
}
\description{
plot_lm plots graphs useful in diagnosing linear models.
}
\details{
This is (sort of) a copy of the stats plot.lm function,
but with ggplot output. It's a work in progress...
}
