% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcheck.R
\name{varcheck}
\alias{varcheck}
\title{Variance check function}
\usage{
varcheck(DT, value, group)
}
\arguments{
\item{DT}{data as data.table (or data frame) object}
}
\description{
Simple function to check within and between group variance
Equivalent to Mean Sq (and F) values from one-way anova
}
